refactor(state): carve WorkflowRunStore + atomic lifecycle ops out of StateDb (#97)
Split the StateDb god-class into three per-table stores sharing one
better-sqlite3 connection, and concentrate run-lifecycle invariants in
named, single-transaction operations.

- ExecutionStore — owns every `executions` query (single owner per table).
- ApprovalStore — `workflow_approvals` CRUD; resolveReplyGate returns its
  row-change count to drive the double-reply concurrency guard.
- WorkflowRunStore — aggregate root, constructed with the injected
  ApprovalStore. Holds the atomic ops, each one db.transaction(...):
  resolveReplyGateAndResume (the socratic-reply hazard), resolveGateAndResume
  (approve), resolveGateAndFail (reject), pauseForApproval (folds
  approval + waiting_approval marker + optional scratch + pause), and
  finishRun({ terminalMarker }) (folds on_success.set_phase). The
  long-running dispatchWorkflow stays outside every transaction — callers
  dispatch after commit.
- StateDb becomes a thin construction root: opens the connection, runs the
  extracted migrate(), builds the stores, exposes db.runs / db.approvals /
  db.executions, re-exports the moved types (zero import churn), and keeps
  only cron-override + kill-switch concerns. Old methods deleted.
- All call sites (index.ts, admin/routes.ts, simple.ts, runner.ts,
  phase-executor.ts, resume.ts, router.ts, scheduler.ts,
  chat-session-reader.ts) rewired; the explore-reply and both
  approval-response flows (GitHub/Slack + dashboard) now call the atomic ops,
  so a throw mid-sequence can't leave a run `running` with its gate
  unresolved or its reply lost.
- phase_history reconciliation deferred (per the issue): the appendPhase seam
  is created now, no behavior change.

`:memory:` is now a genuine per-connection in-memory DB (test isolation) so
parallel test files don't contend on a shared on-disk file once the stores
hold write transactions.

Tests: new workflow-run-store.test.ts drives every atomic op including two
rollback proofs (BigInt poison patch + injected throwing ApprovalStore) and
the double-reply guard; existing state/runner/executor/routes/router tests
migrated to the store APIs. tsc clean, 601/601 tests pass.
