+++ src/state/db.test.ts
@@ -16,7 +16,7 @@ describe("workflow_runs CRUD", () => {
   it("creates a workflow run and retrieves it by ID", () => {
     const id = randomUUID();
     const now = new Date().toISOString();
-    db.createWorkflowRun({
+    db.runs.createRun({
       id,
       workflowName: "build",
       triggerId: "owner/repo#1",
@@ -29,7 +29,7 @@ describe("workflow_runs CRUD", () => {
       finishedAt: undefined,
     });
 
-    const run = db.getWorkflowRun(id);
+    const run = db.runs.getRun(id);
     expect(run).not.toBeNull();
     expect(run!.id).toBe(id);
     expect(run!.workflowName).toBe("build");
@@ -43,13 +43,13 @@ describe("workflow_runs CRUD", () => {
   });
 
   it("returns null for a non-existent ID", () => {
-    expect(db.getWorkflowRun("no-such-id")).toBeNull();
+    expect(db.runs.getRun("no-such-id")).toBeNull();
   });
 
   it("updates phase and appends to phase_history", () => {
     const id = randomUUID();
     const now = new Date().toISOString();
-    db.createWorkflowRun({
+    db.runs.createRun({
       id,
       workflowName: "build",
       triggerId: "owner/repo#2",
@@ -59,9 +59,9 @@ describe("workflow_runs CRUD", () => {
     });
 
     const entry = { phase: "guardrails", timestamp: new Date().toISOString(), success: true, summary: "READY" };
-    db.updateWorkflowPhase(id, "guardrails", entry);
+    db.runs.appendPhase(id, "guardrails", entry);
 
-    const run = db.getWorkflowRun(id);
+    const run = db.runs.getRun(id);
     expect(run!.currentPhase).toBe("guardrails");
     expect(run!.phaseHistory).toHaveLength(1);
     expect(run!.phaseHistory[0]).toEqual(entry);
@@ -70,7 +70,7 @@ describe("workflow_runs CRUD", () => {
   it("appends multiple phase history entries", () => {
     const id = randomUUID();
     const now = new Date().toISOString();
-    db.createWorkflowRun({
+    db.runs.createRun({
       id,
       workflowName: "build",
       triggerId: "owner/repo#3",
@@ -79,11 +79,11 @@ describe("workflow_runs CRUD", () => {
       startedAt: now,
     });
 
-    db.updateWorkflowPhase(id, "guardrails", { phase: "guardrails", timestamp: now, success: true });
-    db.updateWorkflowPhase(id, "architect", { phase: "architect", timestamp: now, success: true });
-    db.updateWorkflowPhase(id, "executor", { phase: "executor", timestamp: now, success: true });
+    db.runs.appendPhase(id, "guardrails", { phase: "guardrails", timestamp: now, success: true });
+    db.runs.appendPhase(id, "architect", { phase: "architect", timestamp: now, success: true });
+    db.runs.appendPhase(id, "executor", { phase: "executor", timestamp: now, success: true });
 
-    const run = db.getWorkflowRun(id);
+    const run = db.runs.getRun(id);
     expect(run!.currentPhase).toBe("executor");
     expect(run!.phaseHistory).toHaveLength(3);
     expect(run!.phaseHistory.map((e) => e.phase)).toEqual(["guardrails", "architect", "executor"]);
@@ -92,7 +92,7 @@ describe("workflow_runs CRUD", () => {
   it("finishes a workflow run with succeeded status", () => {
     const id = randomUUID();
     const now = new Date().toISOString();
-    db.createWorkflowRun({
+    db.runs.createRun({
       id,
       workflowName: "build",
       triggerId: "owner/repo#4",
@@ -101,8 +101,8 @@ describe("workflow_runs CRUD", () => {
       startedAt: now,
     });
 
-    db.finishWorkflowRun(id, "succeeded");
-    const run = db.getWorkflowRun(id);
+    db.runs.finishRun(id, "succeeded");
+    const run = db.runs.getRun(id);
     expect(run!.status).toBe("succeeded");
     expect(run!.finishedAt).toBeTruthy();
   });
@@ -110,7 +110,7 @@ describe("workflow_runs CRUD", () => {
   it("finishes a workflow run with failed status", () => {
     const id = randomUUID();
     const now = new Date().toISOString();
-    db.createWorkflowRun({
+    db.runs.createRun({
       id,
       workflowName: "build",
       triggerId: "owner/repo#5",
@@ -119,8 +119,8 @@ describe("workflow_runs CRUD", () => {
       startedAt: now,
     });
 
-    db.finishWorkflowRun(id, "failed", "some error");
-    const run = db.getWorkflowRun(id);
+    db.runs.finishRun(id, "failed", { error: "some error" });
+    const run = db.runs.getRun(id);
     expect(run!.status).toBe("failed");
     expect(run!.finishedAt).toBeTruthy();
   });
@@ -129,7 +129,7 @@ describe("workflow_runs CRUD", () => {
 describe("node_statuses store removed (issue #94)", () => {
   it("no longer exposes updateNodeStatus and getWorkflowRun has no nodeStatuses", () => {
     const id = randomUUID();
-    db.createWorkflowRun({
+    db.runs.createRun({
       id,
       workflowName: "build",
       triggerId: "owner/repo#77",
@@ -138,7 +138,7 @@ describe("node_statuses store removed (issue #94)", () => {
       startedAt: new Date().toISOString(),
     });
     expect((db as unknown as Record<string, unknown>).updateNodeStatus).toBeUndefined();
-    const run = db.getWorkflowRun(id);
+    const run = db.runs.getRun(id);
     expect(run).not.toBeNull();
     expect((run as unknown as Record<string, unknown>).nodeStatuses).toBeUndefined();
   });
@@ -148,11 +148,11 @@ describe("recordSkippedPhase — skips land in the executions ledger", () => {
   it("writes a finished, non-successful skip row that shouldRunPhase re-evaluates", () => {
     const skill = "build:merge";
     const triggerId = "owner/repo#88";
-    db.recordSkippedPhase(skill, triggerId, "wf-skip-1", "repo");
+    db.executions.recordSkippedPhase(skill, triggerId, "wf-skip-1", "repo");
 
     // Not "done" (success != 1) and not "running" (finished_at set) — so a
     // resume re-evaluates the node (it'll simply be re-skipped if still gated).
-    expect(db.shouldRunPhase(skill, triggerId, "wf-skip-1")).toBe("run");
+    expect(db.executions.shouldRunPhase(skill, triggerId, "wf-skip-1")).toBe("run");
   });
 });
 
@@ -160,7 +160,7 @@ describe("getWorkflowRunByTrigger", () => {
   it("returns the active run for a trigger", () => {
     const id = randomUUID();
     const now = new Date().toISOString();
-    db.createWorkflowRun({
+    db.runs.createRun({
       id,
       workflowName: "build",
       triggerId: "owner/repo#10",
@@ -169,7 +169,7 @@ describe("getWorkflowRunByTrigger", () => {
       startedAt: now,
     });
 
-    const run = db.getWorkflowRunByTrigger("owner/repo#10");
+    const run = db.runs.getByTrigger("owner/repo#10");
     expect(run).not.toBeNull();
     expect(run!.id).toBe(id);
   });
@@ -177,7 +177,7 @@ describe("getWorkflowRunByTrigger", () => {
   it("ignores failed or succeeded runs", () => {
     const id = randomUUID();
     const now = new Date().toISOString();
-    db.createWorkflowRun({
+    db.runs.createRun({
       id,
       workflowName: "build",
       triggerId: "owner/repo#11",
@@ -185,20 +185,20 @@ describe("getWorkflowRunByTrigger", () => {
       status: "running",
       startedAt: now,
     });
-    db.finishWorkflowRun(id, "failed");
+    db.runs.finishRun(id, "failed");
 
-    expect(db.getWorkflowRunByTrigger("owner/repo#11")).toBeNull();
+    expect(db.runs.getByTrigger("owner/repo#11")).toBeNull();
   });
 
   it("returns null when no run exists for trigger", () => {
-    expect(db.getWorkflowRunByTrigger("owner/repo#999")).toBeNull();
+    expect(db.runs.getByTrigger("owner/repo#999")).toBeNull();
   });
 
   it("returns the most recent active run when multiple exist", async () => {
     const id1 = randomUUID();
     const id2 = randomUUID();
 
-    db.createWorkflowRun({
+    db.runs.createRun({
       id: id1,
       workflowName: "build",
       triggerId: "owner/repo#12",
@@ -206,7 +206,7 @@ describe("getWorkflowRunByTrigger", () => {
       status: "running",
       startedAt: new Date(Date.now() - 1000).toISOString(),
     });
-    db.createWorkflowRun({
+    db.runs.createRun({
       id: id2,
       workflowName: "build",
       triggerId: "owner/repo#12",
@@ -215,7 +215,7 @@ describe("getWorkflowRunByTrigger", () => {
       startedAt: new Date().toISOString(),
     });
 
-    const run = db.getWorkflowRunByTrigger("owner/repo#12");
+    const run = db.runs.getByTrigger("owner/repo#12");
     expect(run!.id).toBe(id2);
   });
 });
@@ -225,13 +225,13 @@ describe("activeWorkflowRuns", () => {
     const now = new Date().toISOString();
 
     const runningId = randomUUID();
-    db.createWorkflowRun({ id: runningId, workflowName: "build", triggerId: "t1", currentPhase: "guardrails", status: "running", startedAt: now });
+    db.runs.createRun({ id: runningId, workflowName: "build", triggerId: "t1", currentPhase: "guardrails", status: "running", startedAt: now });
 
     const failedId = randomUUID();
-    db.createWorkflowRun({ id: failedId, workflowName: "build", triggerId: "t2", currentPhase: "executor", status: "running", startedAt: now });
-    db.finishWorkflowRun(failedId, "failed");
+    db.runs.createRun({ id: failedId, workflowName: "build", triggerId: "t2", currentPhase: "executor", status: "running", startedAt: now });
+    db.runs.finishRun(failedId, "failed");
 
-    const active = db.activeWorkflowRuns();
+    const active = db.runs.listActive();
     expect(active.map((r) => r.id)).toContain(runningId);
     expect(active.map((r) => r.id)).not.toContain(failedId);
   });
@@ -241,7 +241,7 @@ describe("recentWorkflowRuns", () => {
   it("respects limit and orders by started_at DESC", () => {
     const now = Date.now();
     for (let i = 0; i < 5; i++) {
-      db.createWorkflowRun({
+      db.runs.createRun({
         id: randomUUID(),
         workflowName: "build",
         triggerId: `t${i}`,
@@ -251,7 +251,7 @@ describe("recentWorkflowRuns", () => {
       });
     }
 
-    const runs = db.recentWorkflowRuns(3);
+    const runs = db.runs.listRecent(3);
     expect(runs).toHaveLength(3);
     // Most recent first
     expect(runs[0]!.startedAt >= runs[1]!.startedAt).toBe(true);
@@ -263,10 +263,10 @@ describe("cancelWorkflowRun", () => {
   it("sets status to cancelled", () => {
     const id = randomUUID();
     const now = new Date().toISOString();
-    db.createWorkflowRun({ id, workflowName: "build", triggerId: "t-cancel", currentPhase: "executor", status: "running", startedAt: now });
-    db.cancelWorkflowRun(id);
+    db.runs.createRun({ id, workflowName: "build", triggerId: "t-cancel", currentPhase: "executor", status: "running", startedAt: now });
+    db.runs.cancelRun(id);
 
-    const run = db.getWorkflowRun(id);
+    const run = db.runs.getRun(id);
     expect(run!.status).toBe("cancelled");
     expect(run!.finishedAt).toBeTruthy();
   });
@@ -275,10 +275,10 @@ describe("cancelWorkflowRun", () => {
 describe("pauseWorkflowRun", () => {
   it("sets status to paused", () => {
     const id = randomUUID();
-    db.createWorkflowRun({ id, workflowName: "build", triggerId: "t-pause", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
-    db.pauseWorkflowRun(id);
+    db.runs.createRun({ id, workflowName: "build", triggerId: "t-pause", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
+    db.runs.setPaused(id);
 
-    const run = db.getWorkflowRun(id);
+    const run = db.runs.getRun(id);
     expect(run!.status).toBe("paused");
   });
 });
@@ -287,10 +287,10 @@ describe("workflow_approvals CRUD", () => {
   it("creates an approval and retrieves it by ID", () => {
     const id = randomUUID();
     const workflowRunId = randomUUID();
-    db.createWorkflowRun({ id: workflowRunId, workflowName: "build", triggerId: "owner/repo#20", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
+    db.runs.createRun({ id: workflowRunId, workflowName: "build", triggerId: "owner/repo#20", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
 
     const now = new Date().toISOString();
-    db.createApproval({
+    db.approvals.create({
       id,
       workflowRunId,
       gate: "post_architect",
@@ -299,7 +299,7 @@ describe("workflow_approvals CRUD", () => {
       createdAt: now,
     });
 
-    const approval = db.getApproval(id);
+    const approval = db.approvals.getById(id);
     expect(approval).not.toBeNull();
     expect(approval!.id).toBe(id);
     expect(approval!.workflowRunId).toBe(workflowRunId);
@@ -311,65 +311,65 @@ describe("workflow_approvals CRUD", () => {
   });
 
   it("returns null for a non-existent approval", () => {
-    expect(db.getApproval("no-such-id")).toBeNull();
+    expect(db.approvals.getById("no-such-id")).toBeNull();
   });
 
   it("getPendingApprovalForWorkflow returns pending approval", () => {
     const workflowRunId = randomUUID();
-    db.createWorkflowRun({ id: workflowRunId, workflowName: "build", triggerId: "owner/repo#21", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
+    db.runs.createRun({ id: workflowRunId, workflowName: "build", triggerId: "owner/repo#21", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
 
     const approvalId = randomUUID();
-    db.createApproval({ id: approvalId, workflowRunId, gate: "post_architect", summary: "Test", createdAt: new Date().toISOString() });
+    db.approvals.create({ id: approvalId, workflowRunId, gate: "post_architect", summary: "Test", createdAt: new Date().toISOString() });
 
-    const approval = db.getPendingApprovalForWorkflow(workflowRunId);
+    const approval = db.approvals.getPendingForWorkflow(workflowRunId);
     expect(approval).not.toBeNull();
     expect(approval!.id).toBe(approvalId);
     expect(approval!.status).toBe("pending");
   });
 
   it("getPendingApprovalForWorkflow returns null when no pending approval", () => {
-    expect(db.getPendingApprovalForWorkflow("no-such-workflow")).toBeNull();
+    expect(db.approvals.getPendingForWorkflow("no-such-workflow")).toBeNull();
   });
 
   it("getPendingApprovalByTrigger returns pending approval by trigger ID", () => {
     const workflowRunId = randomUUID();
     const triggerId = "owner/repo#22";
-    db.createWorkflowRun({ id: workflowRunId, workflowName: "build", triggerId, currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
+    db.runs.createRun({ id: workflowRunId, workflowName: "build", triggerId, currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
 
     const approvalId = randomUUID();
-    db.createApproval({ id: approvalId, workflowRunId, gate: "post_architect", summary: "Test", createdAt: new Date().toISOString() });
+    db.approvals.create({ id: approvalId, workflowRunId, gate: "post_architect", summary: "Test", createdAt: new Date().toISOString() });
 
-    const approval = db.getPendingApprovalByTrigger(triggerId);
+    const approval = db.approvals.getPendingByTrigger(triggerId);
     expect(approval).not.toBeNull();
     expect(approval!.id).toBe(approvalId);
   });
 
   it("getPendingApprovalByTrigger returns null when trigger has no pending approval", () => {
-    expect(db.getPendingApprovalByTrigger("owner/repo#9999")).toBeNull();
+    expect(db.approvals.getPendingByTrigger("owner/repo#9999")).toBeNull();
   });
 
   it("listPendingApprovals returns all pending approvals", () => {
     for (let i = 0; i < 3; i++) {
       const workflowRunId = randomUUID();
-      db.createWorkflowRun({ id: workflowRunId, workflowName: "build", triggerId: `owner/repo#${30 + i}`, currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
-      db.createApproval({ id: randomUUID(), workflowRunId, gate: "post_architect", summary: `Summary ${i}`, createdAt: new Date().toISOString() });
+      db.runs.createRun({ id: workflowRunId, workflowName: "build", triggerId: `owner/repo#${30 + i}`, currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
+      db.approvals.create({ id: randomUUID(), workflowRunId, gate: "post_architect", summary: `Summary ${i}`, createdAt: new Date().toISOString() });
     }
 
-    const pending = db.listPendingApprovals();
+    const pending = db.approvals.listPending();
     expect(pending.length).toBeGreaterThanOrEqual(3);
     expect(pending.every((a) => a.status === "pending")).toBe(true);
   });
 
   it("respondToApproval sets status and respondedBy", () => {
     const workflowRunId = randomUUID();
-    db.createWorkflowRun({ id: workflowRunId, workflowName: "build", triggerId: "owner/repo#23", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
+    db.runs.createRun({ id: workflowRunId, workflowName: "build", triggerId: "owner/repo#23", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
 
     const approvalId = randomUUID();
-    db.createApproval({ id: approvalId, workflowRunId, gate: "post_architect", summary: "Test", createdAt: new Date().toISOString() });
+    db.approvals.create({ id: approvalId, workflowRunId, gate: "post_architect", summary: "Test", createdAt: new Date().toISOString() });
 
-    db.respondToApproval(approvalId, "approved", "bob");
+    db.approvals.respond(approvalId, "approved", "bob");
 
-    const approval = db.getApproval(approvalId);
+    const approval = db.approvals.getById(approvalId);
     expect(approval!.status).toBe("approved");
     expect(approval!.respondedBy).toBe("bob");
     expect(approval!.respondedAt).toBeTruthy();
@@ -378,14 +378,14 @@ describe("workflow_approvals CRUD", () => {
 
   it("respondToApproval stores rejection reason", () => {
     const workflowRunId = randomUUID();
-    db.createWorkflowRun({ id: workflowRunId, workflowName: "build", triggerId: "owner/repo#24", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
+    db.runs.createRun({ id: workflowRunId, workflowName: "build", triggerId: "owner/repo#24", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
 
     const approvalId = randomUUID();
-    db.createApproval({ id: approvalId, workflowRunId, gate: "post_architect", summary: "Test", createdAt: new Date().toISOString() });
+    db.approvals.create({ id: approvalId, workflowRunId, gate: "post_architect", summary: "Test", createdAt: new Date().toISOString() });
 
-    db.respondToApproval(approvalId, "rejected", "carol", "Plan is incomplete");
+    db.approvals.respond(approvalId, "rejected", "carol", "Plan is incomplete");
 
-    const approval = db.getApproval(approvalId);
+    const approval = db.approvals.getById(approvalId);
     expect(approval!.status).toBe("rejected");
     expect(approval!.respondedBy).toBe("carol");
     expect(approval!.response).toBe("Plan is incomplete");
@@ -393,13 +393,13 @@ describe("workflow_approvals CRUD", () => {
 
   it("getPendingApprovalForWorkflow ignores responded approvals", () => {
     const workflowRunId = randomUUID();
-    db.createWorkflowRun({ id: workflowRunId, workflowName: "build", triggerId: "owner/repo#25", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
+    db.runs.createRun({ id: workflowRunId, workflowName: "build", triggerId: "owner/repo#25", currentPhase: "architect", status: "running", startedAt: new Date().toISOString() });
 
     const approvalId = randomUUID();
-    db.createApproval({ id: approvalId, workflowRunId, gate: "post_architect", summary: "Test", createdAt: new Date().toISOString() });
-    db.respondToApproval(approvalId, "approved", "dave");
+    db.approvals.create({ id: approvalId, workflowRunId, gate: "post_architect", summary: "Test", createdAt: new Date().toISOString() });
+    db.approvals.respond(approvalId, "approved", "dave");
 
-    expect(db.getPendingApprovalForWorkflow(workflowRunId)).toBeNull();
+    expect(db.approvals.getPendingForWorkflow(workflowRunId)).toBeNull();
   });
 });
 
@@ -420,7 +420,7 @@ describe("dailyStats", () => {
     cacheReadTokens?: number;
     costUsd?: number;
   }) {
-    db.recordStart({
+    db.executions.recordStart({
       id: opts.id,
       triggerType: "webhook",
       triggerId: "owner/repo#1",
@@ -430,7 +430,7 @@ describe("dailyStats", () => {
       startedAt: opts.startedAt,
     });
     if (opts.success !== undefined) {
-      db.recordFinish(opts.id, {
+      db.executions.recordFinish(opts.id, {
         success: opts.success,
         inputTokens: opts.inputTokens,
         outputTokens: opts.outputTokens,
@@ -451,7 +451,7 @@ describe("dailyStats", () => {
   }
 
   it("returns days rows of zeros when no executions exist", () => {
-    const rows = db.dailyStats(30);
+    const rows = db.executions.dailyStats(30);
     expect(rows).toHaveLength(30);
     expect(rows.every((r) => r.executions === 0)).toBe(true);
     expect(rows.every((r) => r.totalTokens === 0 && r.costUsd === 0)).toBe(true);
@@ -464,7 +464,7 @@ describe("dailyStats", () => {
     insertExecution({ id: randomUUID(), startedAt: day1.iso, success: false });
     insertExecution({ id: randomUUID(), startedAt: day2.iso, success: true });
 
-    const rows = db.dailyStats(30);
+    const rows = db.executions.dailyStats(30);
     expect(rows).toHaveLength(30);
 
     const d1 = rows.find((r) => r.date === day1.key);
@@ -484,7 +484,7 @@ describe("dailyStats", () => {
     insertExecution({ id: randomUUID(), startedAt: day.iso, success: true, inputTokens: 100, outputTokens: 50, cacheReadTokens: 20, costUsd: 0.01 });
     insertExecution({ id: randomUUID(), startedAt: day.iso, success: true, inputTokens: 200, outputTokens: 80, cacheReadTokens: 0, costUsd: 0.02 });
 
-    const rows = db.dailyStats(30);
+    const rows = db.executions.dailyStats(30);
     const d = rows.find((r) => r.date === day.key);
     expect(d).toBeDefined();
     expect(d!.inputTokens).toBe(300);
@@ -497,9 +497,9 @@ describe("dailyStats", () => {
   it("handles NULL token/cost columns gracefully", () => {
     const day = daysAgo(2);
     // recordStart only — no recordFinish, so tokens/cost are NULL
-    db.recordStart({ id: randomUUID(), triggerType: "webhook", triggerId: "t1", skill: "build", repo: "r", issueNumber: 1, startedAt: day.iso });
+    db.executions.recordStart({ id: randomUUID(), triggerType: "webhook", triggerId: "t1", skill: "build", repo: "r", issueNumber: 1, startedAt: day.iso });
 
-    const rows = db.dailyStats(30);
+    const rows = db.executions.dailyStats(30);
     const d = rows.find((r) => r.date === day.key);
     expect(d).toBeDefined();
     expect(d!.totalTokens).toBe(0);
@@ -515,7 +515,7 @@ describe("dailyStats", () => {
     // Recent execution — today
     insertExecution({ id: randomUUID(), startedAt: new Date().toISOString(), success: true });
 
-    const rows = db.dailyStats(30);
+    const rows = db.executions.dailyStats(30);
     // 30 daily rows (filled with zeros), with exactly one having an execution
     expect(rows).toHaveLength(30);
     const withExec = rows.filter((r) => r.executions > 0);
@@ -530,7 +530,7 @@ describe("dailyStats", () => {
     insertExecution({ id: randomUUID(), startedAt: d1, success: true });
     insertExecution({ id: randomUUID(), startedAt: d2, success: true });
 
-    const rows = db.dailyStats(30);
+    const rows = db.executions.dailyStats(30);
     expect(rows).toHaveLength(30);
     const dates = rows.map((r) => r.date);
     const sorted = [...dates].sort();
@@ -542,7 +542,7 @@ describe("context JSON round-trip", () => {
   it("stores and retrieves complex context objects", () => {
     const id = randomUUID();
     const context = { branch: "lastlight/42-my-feature", taskId: "repo-42", models: { architect: "claude-opus-4-6" } };
-    db.createWorkflowRun({
+    db.runs.createRun({
       id,
       workflowName: "build",
       triggerId: "owner/repo#42",
@@ -552,13 +552,13 @@ describe("context JSON round-trip", () => {
       startedAt: new Date().toISOString(),
     });
 
-    const run = db.getWorkflowRun(id);
+    const run = db.runs.getRun(id);
     expect(run!.context).toEqual(context);
   });
 
   it("handles runs without context", () => {
     const id = randomUUID();
-    db.createWorkflowRun({
+    db.runs.createRun({
       id,
       workflowName: "build",
       triggerId: "owner/repo#43",
@@ -567,7 +567,7 @@ describe("context JSON round-trip", () => {
       startedAt: new Date().toISOString(),
     });
 
-    const run = db.getWorkflowRun(id);
+    const run = db.runs.getRun(id);
     expect(run!.context).toBeUndefined();
   });
 });
