+++ src/state/db.ts
@@ -1,63 +1,23 @@
 import Database from "better-sqlite3";
 import { resolve } from "path";
-import { randomUUID } from "crypto";
+import { migrate } from "./migrate.js";
+import { ExecutionStore } from "./execution-store.js";
+import { ApprovalStore } from "./approval-store.js";
+import { WorkflowRunStore } from "./workflow-run-store.js";
+
+// Re-export the types that moved out to the per-table stores so existing
+// import sites (`import { WorkflowRun } from "../state/db.js"`, etc.) keep
+// working unchanged. The carve-out (issue #97) splits the implementation
+// across three stores but preserves db.ts as the single import surface for
+// the shared vocabulary.
+export type { ExecutionRecord } from "./execution-store.js";
+export type { WorkflowApproval } from "./approval-store.js";
+export type { WorkflowRun, PhaseHistoryEntry, PhaseMarker } from "./workflow-run-store.js";
+export { ExecutionStore } from "./execution-store.js";
+export { ApprovalStore } from "./approval-store.js";
+export { WorkflowRunStore } from "./workflow-run-store.js";
 
 const DEFAULT_DB_PATH = "lastlight.db";
-
-export interface PhaseHistoryEntry {
-  phase: string;
-  timestamp: string;
-  success: boolean;
-  summary?: string;
-}
-
-export interface WorkflowApproval {
-  id: string;
-  workflowRunId: string;
-  gate: string;
-  summary: string;
-  status: 'pending' | 'approved' | 'rejected';
-  /**
-   * Gate flavor. `approve` gates resolve only on an explicit approve/reject
-   * command; `reply` gates resolve on any free-form reply in the same
-   * thread (used by the socratic explore loop).
-   */
-  kind: 'approve' | 'reply';
-  requestedBy?: string;
-  respondedBy?: string;
-  response?: string;
-  respondedAt?: string;
-  createdAt: string;
-}
-
-export interface WorkflowRun {
-  id: string;
-  workflowName: string;
-  triggerId: string;
-  repo?: string;
-  issueNumber?: number;
-  currentPhase: string;
-  phaseHistory: PhaseHistoryEntry[];
-  status: "running" | "paused" | "succeeded" | "failed" | "cancelled";
-  context?: Record<string, unknown>;
-  /**
-   * Mutable phase-to-phase state, merged at the top level by
-   * `updateWorkflowRunScratch`. Distinct from `context` (which is the
-   * immutable trigger input) — used by features like the socratic explore
-   * loop to accumulate Q&A across reply-gate pauses.
-   */
-  scratch?: Record<string, unknown>;
-  /**
-   * Number of times `resumeOrphanedWorkflows` has re-dispatched this run
-   * after a harness restart. Bounded by a small limit so a run that
-   * crashes the host (e.g. agent OOM) is marked failed instead of
-   * retried forever.
-   */
-  restartCount?: number;
-  startedAt: string;
-  updatedAt: string;
-  finishedAt?: string;
-}
 
 export interface CronOverride {
   name: string;
@@ -81,1185 +41,42 @@ export interface WorkflowOverride {
   updatedBy: string | null;
 }
 
-export interface ExecutionRecord {
-  id: string;
-  triggerType: "webhook" | "cron" | "chat" | "api";
-  triggerId: string;
-  skill: string;
-  repo?: string;
-  issueNumber?: number;
-  startedAt: string;
-  finishedAt?: string;
-  success?: boolean;
-  error?: string;
-  turns?: number;
-  durationMs?: number;
-  /**
-   * Runtime session id (OpenCode `ses_…`) captured from the event stream.
-   * Used by the dashboard to look up the session log for a given phase.
-   */
-  sessionId?: string;
-  /** Total USD cost reported by the runtime on the final step. Zero under OAuth/subscription auth. */
-  costUsd?: number;
-  /** Fresh input tokens (no cache hit). */
-  inputTokens?: number;
-  /** Tokens that triggered a cache write. */
-  cacheCreationInputTokens?: number;
-  /** Tokens served from prompt cache. */
-  cacheReadInputTokens?: number;
-  /** Output tokens generated by the model. */
-  outputTokens?: number;
-  /** API-side duration (excludes orchestrator overhead). */
-  apiDurationMs?: number;
-  /** Result subtype, e.g. "success" or "error_max_turns". */
-  stopReason?: string;
-  /**
-   * JSON-serialized {@link ExtensionStatusMap} — which agentic-pi extensions
-   * (file-search / github / web-search) were active for this execution.
-   */
-  extensionStatus?: string;
-  /**
-   * Workflow run that owns this execution. Scopes dedup: a fresh re-trigger
-   * creates a new workflow_run_id, so `shouldRunPhase` won't find matching
-   * rows from the prior run and will correctly run the phase again.
-   */
-  workflowRunId?: string;
-}
-
 /**
  * Lightweight SQLite state for operational tracking.
  * NOT for conversation history — only execution logs and rate limits.
+ *
+ * Construction root for the three per-table stores (issue #97): it opens the
+ * single `Database` connection, runs {@link migrate}, and builds
+ * {@link ExecutionStore} / {@link ApprovalStore} / {@link WorkflowRunStore} on
+ * top of it (all sharing the one connection so better-sqlite3 transactions
+ * span every store). Callers reach the stores via `db.runs` / `db.approvals` /
+ * `db.executions`. StateDb itself retains only the cross-cutting cron-override
+ * and workflow kill-switch concerns.
  */
 export class StateDb {
   private db: Database.Database;
 
+  /** Owns the `executions` ledger. */
+  readonly executions: ExecutionStore;
+  /** Owns the `workflow_approvals` table. */
+  readonly approvals: ApprovalStore;
+  /** Aggregate root for `workflow_runs` + the atomic lifecycle operations. */
+  readonly runs: WorkflowRunStore;
+
   constructor(dbPath?: string) {
-    this.db = new Database(resolve(dbPath || DEFAULT_DB_PATH));
+    // ":memory:" stays a real per-connection in-memory DB (used by tests for
+    // isolation) — only filesystem paths go through resolve(). Previously
+    // resolve(":memory:") produced a shared on-disk file named ":memory:",
+    // which let parallel test files contend on one DB and deadlock once the
+    // stores started holding write transactions.
+    const target = dbPath === ":memory:" ? ":memory:" : resolve(dbPath || DEFAULT_DB_PATH);
+    this.db = new Database(target);
     this.db.pragma("journal_mode = WAL");
-    this.migrate();
-  }
-
-  private migrate() {
-    this.db.exec(`
-      CREATE TABLE IF NOT EXISTS executions (
-        id TEXT PRIMARY KEY,
-        trigger_type TEXT NOT NULL,
-        trigger_id TEXT NOT NULL,
-        skill TEXT NOT NULL,
-        repo TEXT,
-        issue_number INTEGER,
-        started_at TEXT NOT NULL,
-        finished_at TEXT,
-        success INTEGER,
-        error TEXT,
-        turns INTEGER,
-        duration_ms INTEGER
-      );
-
-      CREATE INDEX IF NOT EXISTS idx_executions_trigger ON executions(trigger_type, trigger_id);
-      CREATE INDEX IF NOT EXISTS idx_executions_skill ON executions(skill, started_at);
-
-      CREATE TABLE IF NOT EXISTS workflow_runs (
-        id TEXT PRIMARY KEY,
-        workflow_name TEXT NOT NULL,
-        trigger_id TEXT NOT NULL,
-        repo TEXT,
-        issue_number INTEGER,
-        current_phase TEXT NOT NULL,
-        phase_history TEXT NOT NULL DEFAULT '[]',
-        status TEXT NOT NULL DEFAULT 'running',
-        context TEXT,
-        started_at TEXT NOT NULL,
-        updated_at TEXT NOT NULL,
-        finished_at TEXT
-      );
-      CREATE INDEX IF NOT EXISTS idx_workflow_runs_trigger ON workflow_runs(trigger_id, status);
-      CREATE INDEX IF NOT EXISTS idx_workflow_runs_status ON workflow_runs(status);
-      -- Dashboard's workflow-runs list query is ORDER BY started_at DESC
-      -- LIMIT 20, often with optional workflow_name / status filters and a
-      -- companion COUNT(*). It polls every 5s, so it's worth a covering
-      -- index — without one this is a full-table sort.
-      CREATE INDEX IF NOT EXISTS idx_workflow_runs_started_at ON workflow_runs(started_at DESC);
-      CREATE INDEX IF NOT EXISTS idx_workflow_runs_name_started ON workflow_runs(workflow_name, started_at DESC);
-
-      CREATE TABLE IF NOT EXISTS cron_overrides (
-        name TEXT PRIMARY KEY,
-        enabled INTEGER NOT NULL DEFAULT 1,
-        schedule TEXT,
-        updated_at TEXT NOT NULL,
-        updated_by TEXT
-      );
-
-      CREATE TABLE IF NOT EXISTS workflow_overrides (
-        name TEXT PRIMARY KEY,
-        enabled INTEGER NOT NULL DEFAULT 1,
-        updated_at TEXT NOT NULL,
-        updated_by TEXT
-      );
-
-      CREATE TABLE IF NOT EXISTS workflow_approvals (
-        id TEXT PRIMARY KEY,
-        workflow_run_id TEXT NOT NULL,
-        gate TEXT NOT NULL,
-        summary TEXT NOT NULL,
-        status TEXT NOT NULL DEFAULT 'pending',
-        requested_by TEXT,
-        responded_by TEXT,
-        response TEXT,
-        responded_at TEXT,
-        created_at TEXT NOT NULL
-      );
-      CREATE INDEX IF NOT EXISTS idx_approvals_workflow ON workflow_approvals(workflow_run_id);
-      CREATE INDEX IF NOT EXISTS idx_approvals_status ON workflow_approvals(status);
-    `);
-
-    // Mutable phase-to-phase state for features like the socratic explore
-    // loop that accumulate data across reply-gate pauses.
-    try {
-      this.db.exec(`ALTER TABLE workflow_runs ADD COLUMN scratch TEXT`);
-    } catch {
-      // Column already exists — ignore
-    }
-
-    // Restart attempt counter — bumped each time resumeOrphanedWorkflows
-    // re-dispatches a still-'running' row at boot. Acts as a circuit
-    // breaker so an OOM-on-restart loop can't churn forever.
-    try {
-      this.db.exec(`ALTER TABLE workflow_runs ADD COLUMN restart_count INTEGER NOT NULL DEFAULT 0`);
-    } catch {
-      // Column already exists — ignore
-    }
-
-    // Gate flavor: 'approve' (explicit approve/reject) vs 'reply' (resolves
-    // on any free-form reply in the same thread — used by the socratic
-    // explore loop).
-    try {
-      this.db.exec(`ALTER TABLE workflow_approvals ADD COLUMN kind TEXT NOT NULL DEFAULT 'approve'`);
-    } catch {
-      // Column already exists — ignore
-    }
-
-    // Add session_id column to executions so the dashboard can resolve a
-    // phase click to its agent session log. Additive — safe on existing DBs.
-    try {
-      this.db.exec(`ALTER TABLE executions ADD COLUMN session_id TEXT`);
-    } catch {
-      // Column already exists — ignore
-    }
-
-    // Per-execution runtime usage metrics, captured from the OpenCode
-    // step_finish events. Lets the dashboard show cost / tokens per phase
-    // and lets us aggregate spend later. All additive, safe on existing DBs.
-    for (const col of [
-      "cost_usd REAL",
-      "input_tokens INTEGER",
-      "cache_creation_input_tokens INTEGER",
-      "cache_read_input_tokens INTEGER",
-      "output_tokens INTEGER",
-      "api_duration_ms INTEGER",
-      "stop_reason TEXT",
-      // Links a phase execution to its owning workflow_run. Dedup and
-      // "already running / done" checks scope to this column so a fresh
-      // re-trigger of the same workflow on the same issue creates new
-      // executions instead of reusing the old run's rows.
-      "workflow_run_id TEXT",
-      // Final assistant text for a phase execution. Populated only for
-      // loop iterations whose output is referenced by `scratch.<key>
-      // .lastOutputExecutionId` — keeps the inlined LLM output out of
-      // `workflow_runs.scratch` (which is read on every list query)
-      // while still being available to the runner's resume path.
-      "output_text TEXT",
-      // JSON map of agentic-pi extensions active for this execution
-      // (file-search / github / web-search → {status, mode?, provider?,
-      // toolCount?, reason?}). Surfaced in the dashboard phase-detail panel.
-      "extension_status TEXT",
-    ]) {
-      try {
-        this.db.exec(`ALTER TABLE executions ADD COLUMN ${col}`);
-      } catch {
-        // Column already exists — ignore
-      }
-    }
-    try {
-      this.db.exec(
-        `CREATE INDEX IF NOT EXISTS idx_executions_workflow_run ON executions(workflow_run_id, skill)`,
-      );
-    } catch {
-      // Index already exists — ignore
-    }
-  }
-
-  recordStart(record: Omit<ExecutionRecord, "finishedAt" | "success" | "error" | "turns" | "durationMs">): void {
-    this.db.prepare(`
-      INSERT INTO executions (id, trigger_type, trigger_id, skill, repo, issue_number, started_at, workflow_run_id)
-      VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-    `).run(
-      record.id,
-      record.triggerType,
-      record.triggerId,
-      record.skill,
-      record.repo,
-      record.issueNumber,
-      record.startedAt,
-      record.workflowRunId ?? null,
-    );
-  }
-
-  /**
-   * Mid-run update: persist the agent session id as soon as the executor sees
-   * the stream-json `system/init` line. Lets the dashboard surface the live
-   * session log for an in-flight phase, not just for completed ones.
-   */
-  recordSessionId(id: string, sessionId: string): void {
-    this.db.prepare(`UPDATE executions SET session_id = ? WHERE id = ?`).run(sessionId, id);
-  }
-
-  /**
-   * Persist the final assistant text for a loop-iteration execution.
-   * Used by the runner when a scratch slot would otherwise inline the
-   * full LLM output into `workflow_runs.scratch` — the scratch entry
-   * stores `lastOutputExecutionId` and resolves the text through
-   * `getExecutionOutput` on read.
-   */
-  recordOutputText(id: string, output: string): void {
-    this.db.prepare(`UPDATE executions SET output_text = ? WHERE id = ?`).run(output, id);
-  }
-
-  /** Read the persisted output text for an execution, or null if none. */
-  getExecutionOutput(id: string): string | null {
-    const row = this.db
-      .prepare(`SELECT output_text FROM executions WHERE id = ?`)
-      .get(id) as { output_text: string | null } | undefined;
-    return row?.output_text ?? null;
-  }
-
-  /**
-   * Most recent finished execution's `output_text` for a phase (`skill`), or
-   * null. Used by the reviewer loop on resume to re-derive a completed review's
-   * verdict from its persisted output (a dedup-`done` review must not be
-   * assumed APPROVED — it may have requested changes).
-   */
-  getPhaseOutput(skill: string, triggerId: string, workflowRunId?: string): string | null {
-    const [scopeClause, scopeParam] = workflowRunId
-      ? ["workflow_run_id = ?", workflowRunId]
-      : ["trigger_id = ?", triggerId];
-    const row = this.db.prepare(`
-      SELECT output_text FROM executions
-      WHERE skill = ? AND ${scopeClause} AND output_text IS NOT NULL
-      ORDER BY finished_at DESC
-      LIMIT 1
-    `).get(skill, scopeParam) as { output_text: string | null } | undefined;
-    return row?.output_text ?? null;
-  }
-
-  recordFinish(
-    id: string,
-    result: {
-      success: boolean;
-      error?: string;
-      turns?: number;
-      durationMs?: number;
-      sessionId?: string;
-      costUsd?: number;
-      inputTokens?: number;
-      cacheCreationInputTokens?: number;
-      cacheReadInputTokens?: number;
-      outputTokens?: number;
-      apiDurationMs?: number;
-      stopReason?: string;
-      /** JSON-serialized ExtensionStatusMap (extensions active this run). */
-      extensionStatus?: string;
-    },
-  ): void {
-    this.db.prepare(`
-      UPDATE executions
-      SET finished_at = ?,
-          success = ?,
-          error = ?,
-          turns = ?,
-          duration_ms = ?,
-          session_id = COALESCE(?, session_id),
-          cost_usd = COALESCE(?, cost_usd),
-          input_tokens = COALESCE(?, input_tokens),
-          cache_creation_input_tokens = COALESCE(?, cache_creation_input_tokens),
-          cache_read_input_tokens = COALESCE(?, cache_read_input_tokens),
-          output_tokens = COALESCE(?, output_tokens),
-          api_duration_ms = COALESCE(?, api_duration_ms),
-          stop_reason = COALESCE(?, stop_reason),
-          extension_status = COALESCE(?, extension_status)
-      WHERE id = ?
-    `).run(
-      new Date().toISOString(),
-      result.success ? 1 : 0,
-      result.error,
-      result.turns,
-      result.durationMs,
-      result.sessionId ?? null,
-      result.costUsd ?? null,
-      result.inputTokens ?? null,
-      result.cacheCreationInputTokens ?? null,
-      result.cacheReadInputTokens ?? null,
-      result.outputTokens ?? null,
-      result.apiDurationMs ?? null,
-      result.stopReason ?? null,
-      result.extensionStatus ?? null,
-      id,
-    );
-  }
-
-  /**
-   * Record a phase that the scheduler skipped because its trigger rule was
-   * not satisfied (a failed/skipped upstream cascaded down a chain). The
-   * executions ledger is the single source of truth the dashboard derives
-   * phase status from, so skips are written here too — as a finished,
-   * non-successful row (`success = 0`, `stop_reason = 'skipped'`). Because it
-   * is not `success = 1`, `shouldRunPhase` re-evaluates it on resume (the node
-   * will simply be re-skipped if the upstream is still failed).
-   */
-  recordSkippedPhase(
-    skill: string,
-    triggerId: string,
-    workflowRunId?: string,
-    repo?: string,
-  ): void {
-    const now = new Date().toISOString();
-    const m = triggerId.match(/#(\d+)$/);
-    const issueNumber = m ? Number(m[1]) : null;
-    this.db.prepare(`
-      INSERT INTO executions
-        (id, trigger_type, trigger_id, skill, repo, issue_number, started_at, finished_at, success, error, stop_reason, workflow_run_id)
-      VALUES (?, 'webhook', ?, ?, ?, ?, ?, ?, 0, 'skipped: trigger rule not satisfied', 'skipped', ?)
-    `).run(
-      randomUUID(),
-      triggerId,
-      skill,
-      repo ?? null,
-      issueNumber,
-      now,
-      now,
-      workflowRunId ?? null,
-    );
-  }
-
-  /**
-   * List chat threads aggregated from the executions table.
-   *
-   * Each row collapses every execution sharing a `trigger_id` (the messaging
-   * session id) into one summary so the dashboard's chat-sessions tab can
-   * show one card per Slack thread, not one per turn. Joined with
-   * `messaging_sessions` for the agent SDK session id (used to resolve to
-   * the on-disk jsonl when streaming messages) and with `messaging_messages`
-   * for the most recent assistant text (preview snippet).
-   */
-  listChatThreads(limit: number): {
-    triggerId: string;
-    agentSessionId: string | null;
-    firstStartedAt: string;
-    lastActivityAt: string;
-    turnCount: number;
-    totalCost: number;
-    inputTokens: number;
-    outputTokens: number;
-    cacheReadTokens: number;
-    lastAssistantContent: string | null;
-  }[] {
-    const rows = this.db.prepare(`
-      SELECT
-        e.trigger_id              AS triggerId,
-        ms.agent_session_id       AS agentSessionId,
-        MIN(e.started_at)         AS firstStartedAt,
-        MAX(COALESCE(e.finished_at, e.started_at)) AS lastActivityAt,
-        COUNT(*)                  AS turnCount,
-        SUM(COALESCE(e.cost_usd, 0))           AS totalCost,
-        SUM(COALESCE(e.input_tokens, 0))       AS inputTokens,
-        SUM(COALESCE(e.output_tokens, 0))      AS outputTokens,
-        SUM(COALESCE(e.cache_read_input_tokens, 0)) AS cacheReadTokens,
-        (
-          SELECT mm.content
-          FROM messaging_messages mm
-          WHERE mm.session_id = e.trigger_id AND mm.role = 'assistant'
-          ORDER BY mm.timestamp DESC
-          LIMIT 1
-        ) AS lastAssistantContent
-      FROM executions e
-      LEFT JOIN messaging_sessions ms ON ms.id = e.trigger_id
-      WHERE e.skill = 'chat'
-      GROUP BY e.trigger_id
-      ORDER BY lastActivityAt DESC
-      LIMIT ?
-    `).all(limit) as Array<{
-      triggerId: string;
-      agentSessionId: string | null;
-      firstStartedAt: string;
-      lastActivityAt: string;
-      turnCount: number;
-      totalCost: number;
-      inputTokens: number;
-      outputTokens: number;
-      cacheReadTokens: number;
-      lastAssistantContent: string | null;
-    }>;
-    return rows;
-  }
-
-  /** Look up a single chat thread by its messaging session id (= trigger_id). */
-  getChatThread(triggerId: string): {
-    triggerId: string;
-    agentSessionId: string | null;
-    firstStartedAt: string;
-    lastActivityAt: string;
-    turnCount: number;
-    totalCost: number;
-    inputTokens: number;
-    outputTokens: number;
-    cacheReadTokens: number;
-    lastAssistantContent: string | null;
-  } | null {
-    const row = this.db.prepare(`
-      SELECT
-        e.trigger_id              AS triggerId,
-        ms.agent_session_id       AS agentSessionId,
-        MIN(e.started_at)         AS firstStartedAt,
-        MAX(COALESCE(e.finished_at, e.started_at)) AS lastActivityAt,
-        COUNT(*)                  AS turnCount,
-        SUM(COALESCE(e.cost_usd, 0))           AS totalCost,
-        SUM(COALESCE(e.input_tokens, 0))       AS inputTokens,
-        SUM(COALESCE(e.output_tokens, 0))      AS outputTokens,
-        SUM(COALESCE(e.cache_read_input_tokens, 0)) AS cacheReadTokens,
-        (
-          SELECT mm.content
-          FROM messaging_messages mm
-          WHERE mm.session_id = e.trigger_id AND mm.role = 'assistant'
-          ORDER BY mm.timestamp DESC
-          LIMIT 1
-        ) AS lastAssistantContent
-      FROM executions e
-      LEFT JOIN messaging_sessions ms ON ms.id = e.trigger_id
-      WHERE e.skill = 'chat' AND e.trigger_id = ?
-      GROUP BY e.trigger_id
-    `).get(triggerId) as {
-      triggerId: string;
-      agentSessionId: string | null;
-      firstStartedAt: string;
-      lastActivityAt: string;
-      turnCount: number;
-      totalCost: number;
-      inputTokens: number;
-      outputTokens: number;
-      cacheReadTokens: number;
-      lastAssistantContent: string | null;
-    } | undefined;
-    return row ?? null;
-  }
-
-  /** Check if a skill is currently running for a given trigger */
-  isRunning(skill: string, triggerId: string): boolean {
-    const row = this.db.prepare(`
-      SELECT 1 FROM executions
-      WHERE skill = ? AND trigger_id = ? AND finished_at IS NULL
-      LIMIT 1
-    `).get(skill, triggerId);
-    return !!row;
-  }
-
-  /** Check if a skill has already completed successfully for a given trigger */
-  isCompleted(skill: string, triggerId: string): boolean {
-    const row = this.db.prepare(`
-      SELECT 1 FROM executions
-      WHERE skill = ? AND trigger_id = ? AND success = 1
-      LIMIT 1
-    `).get(skill, triggerId);
-    return !!row;
-  }
-
-  /**
-   * Check if a phase should run: not currently running AND not already
-   * succeeded. When `workflowRunId` is provided, scopes the check to that
-   * specific run — a fresh re-trigger (new workflow_run_id) will see no
-   * matching rows and correctly return "run" even if a prior completed run
-   * for the same trigger exists. Omit `workflowRunId` for legacy callers
-   * (e.g. chat/status-report dedup) that want the old trigger-wide scope.
-   */
-  shouldRunPhase(skill: string, triggerId: string, workflowRunId?: string): "run" | "running" | "done" {
-    const [scopeClause, scopeParam] = workflowRunId
-      ? ["workflow_run_id = ?", workflowRunId]
-      : ["trigger_id = ?", triggerId];
-
-    const running = this.db.prepare(`
-      SELECT 1 FROM executions
-      WHERE skill = ? AND ${scopeClause} AND finished_at IS NULL
-      LIMIT 1
-    `).get(skill, scopeParam);
-    if (running) return "running";
-
-    const done = this.db.prepare(`
-      SELECT 1 FROM executions
-      WHERE skill = ? AND ${scopeClause} AND success = 1
-      LIMIT 1
-    `).get(skill, scopeParam);
-    if (done) return "done";
-
-    return "run";
-  }
-
-  /** Mark all stale "running" executions for a skill/trigger as failed.
-   *  Called when we detect no matching Docker container is alive. */
-  markStaleAsFailed(skill: string, triggerId: string, workflowRunId?: string): number {
-    const [scopeClause, scopeParam] = workflowRunId
-      ? ["workflow_run_id = ?", workflowRunId]
-      : ["trigger_id = ?", triggerId];
-    const result = this.db.prepare(`
-      UPDATE executions
-      SET finished_at = ?, success = 0, error = 'stale: container no longer running'
-      WHERE skill = ? AND ${scopeClause} AND finished_at IS NULL
-    `).run(new Date().toISOString(), skill, scopeParam);
-    return result.changes;
-  }
-
-  /**
-   * Mark every unfinished execution row for a trigger as failed, regardless
-   * of skill. Used by the boot-time recovery sweep to clear out phases that
-   * were running when the harness crashed/restarted, so the runner's dedup
-   * logic doesn't think they're still in flight.
-   */
-  markAllStaleForTrigger(triggerId: string, reason: string): number {
-    const result = this.db.prepare(`
-      UPDATE executions
-      SET finished_at = ?, success = 0, error = ?
-      WHERE trigger_id = ? AND finished_at IS NULL
-    `).run(new Date().toISOString(), reason, triggerId);
-    return result.changes;
-  }
-
-  /**
-   * Mark the most recent execution for a skill/trigger as failed.
-   * Used when the agent SDK reported success but the orchestrator decided
-   * the business outcome was a failure (e.g. guardrails BLOCKED, reviewer
-   * REQUEST_CHANGES). Without this, runPhase would skip the phase on the
-   * next run because the DB still says success=1.
-   */
-  markLatestAsFailed(skill: string, triggerId: string, reason: string, workflowRunId?: string): number {
-    const [scopeClause, scopeParam] = workflowRunId
-      ? ["workflow_run_id = ?", workflowRunId]
-      : ["trigger_id = ?", triggerId];
-    const result = this.db.prepare(`
-      UPDATE executions
-      SET success = 0, error = ?
-      WHERE id = (
-        SELECT id FROM executions
-        WHERE skill = ? AND ${scopeClause}
-        ORDER BY started_at DESC
-        LIMIT 1
-      )
-    `).run(reason, skill, scopeParam);
-    return result.changes;
-  }
-
-  /** Get recent executions for a skill */
-  recentExecutions(skill: string, limit = 10): ExecutionRecord[] {
-    return this.db.prepare(`
-      SELECT * FROM executions
-      WHERE skill = ?
-      ORDER BY started_at DESC
-      LIMIT ?
-    `).all(skill, limit) as ExecutionRecord[];
-  }
-
-  /** Count consecutive failures for a skill (for cron failure tracking) */
-  consecutiveFailures(skill: string): number {
-    const rows = this.db.prepare(`
-      SELECT success FROM executions
-      WHERE skill = ? AND finished_at IS NOT NULL
-      ORDER BY started_at DESC
-      LIMIT 10
-    `).all(skill) as { success: number }[];
-
-    let count = 0;
-    for (const row of rows) {
-      if (row.success === 0) count++;
-      else break;
-    }
-    return count;
-  }
-
-  /** Get all executions with pagination */
-  allExecutions(limit = 100, offset = 0): ExecutionRecord[] {
-    return this.db.prepare(`
-      SELECT * FROM executions
-      ORDER BY started_at DESC
-      LIMIT ? OFFSET ?
-    `).all(limit, offset) as ExecutionRecord[];
-  }
-
-  /**
-   * Get every execution recorded for a workflow run, with camelCase fields
-   * ready for the dashboard. Ordered by start time ascending so phases
-   * display in the order they ran. Scopes to a specific `workflow_run_id`
-   * when provided (preferred — isolates fresh re-triggers from prior runs);
-   * falls back to the legacy trigger-wide scope for rows written before the
-   * workflow_run_id column existed.
-   */
-  getExecutionsForWorkflowRun(workflowRunId: string, triggerId: string, workflowName?: string): ExecutionRecord[] {
-    const skillPattern = workflowName ? `${workflowName}:%` : "%:%";
-    const rows = this.db.prepare(`
-      SELECT
-        id,
-        trigger_type    AS triggerType,
-        trigger_id      AS triggerId,
-        skill,
-        repo,
-        issue_number    AS issueNumber,
-        started_at      AS startedAt,
-        finished_at     AS finishedAt,
-        success,
-        error,
-        turns,
-        duration_ms     AS durationMs,
-        session_id      AS sessionId,
-        cost_usd        AS costUsd,
-        input_tokens    AS inputTokens,
-        cache_creation_input_tokens AS cacheCreationInputTokens,
-        cache_read_input_tokens     AS cacheReadInputTokens,
-        output_tokens   AS outputTokens,
-        api_duration_ms AS apiDurationMs,
-        stop_reason     AS stopReason,
-        extension_status AS extensionStatus,
-        workflow_run_id AS workflowRunId
-      FROM executions
-      WHERE (workflow_run_id = ? OR (workflow_run_id IS NULL AND trigger_id = ?))
-        AND skill LIKE ?
-      ORDER BY started_at ASC
-    `).all(workflowRunId, triggerId, skillPattern) as Array<Record<string, unknown>>;
-
-    return rows.map((r) => ({
-      id: r.id as string,
-      triggerType: r.triggerType as ExecutionRecord["triggerType"],
-      triggerId: r.triggerId as string,
-      skill: r.skill as string,
-      repo: (r.repo as string | null) ?? undefined,
-      issueNumber: (r.issueNumber as number | null) ?? undefined,
-      startedAt: r.startedAt as string,
-      finishedAt: (r.finishedAt as string | null) ?? undefined,
-      success: r.success === null || r.success === undefined ? undefined : Boolean(r.success),
-      error: (r.error as string | null) ?? undefined,
-      turns: (r.turns as number | null) ?? undefined,
-      durationMs: (r.durationMs as number | null) ?? undefined,
-      sessionId: (r.sessionId as string | null) ?? undefined,
-      costUsd: (r.costUsd as number | null) ?? undefined,
-      inputTokens: (r.inputTokens as number | null) ?? undefined,
-      cacheCreationInputTokens: (r.cacheCreationInputTokens as number | null) ?? undefined,
-      cacheReadInputTokens: (r.cacheReadInputTokens as number | null) ?? undefined,
-      outputTokens: (r.outputTokens as number | null) ?? undefined,
-      apiDurationMs: (r.apiDurationMs as number | null) ?? undefined,
-      stopReason: (r.stopReason as string | null) ?? undefined,
-      extensionStatus: (r.extensionStatus as string | null) ?? undefined,
-      workflowRunId: (r.workflowRunId as string | null) ?? undefined,
-    }));
-  }
-
-  /** Get currently running executions (no finished_at) */
-  runningExecutions(): ExecutionRecord[] {
-    return this.db.prepare(`
-      SELECT * FROM executions
-      WHERE finished_at IS NULL
-      ORDER BY started_at DESC
-    `).all() as ExecutionRecord[];
-  }
-
-  /** Aggregate execution stats */
-  executionStats(): {
-    total_executions: number;
-    today_count: number;
-    by_skill: Record<string, { count: number; success: number; fail: number }>;
-    by_trigger: Record<string, number>;
-    running: number;
-  } {
-    const today = new Date();
-    today.setHours(0, 0, 0, 0);
-    const todayIso = today.toISOString();
-
-    const total = (this.db.prepare(`SELECT COUNT(*) as c FROM executions`).get() as { c: number }).c;
-    const todayCount = (this.db.prepare(`SELECT COUNT(*) as c FROM executions WHERE started_at >= ?`).get(todayIso) as { c: number }).c;
-    const running = (this.db.prepare(`SELECT COUNT(*) as c FROM executions WHERE finished_at IS NULL`).get() as { c: number }).c;
-
-    const skillRows = this.db.prepare(`
-      SELECT skill, COUNT(*) as count,
-        SUM(CASE WHEN success = 1 THEN 1 ELSE 0 END) as success,
-        SUM(CASE WHEN success = 0 THEN 1 ELSE 0 END) as fail
-      FROM executions GROUP BY skill
-    `).all() as { skill: string; count: number; success: number; fail: number }[];
-
-    const by_skill: Record<string, { count: number; success: number; fail: number }> = {};
-    for (const r of skillRows) {
-      by_skill[r.skill] = { count: r.count, success: r.success, fail: r.fail };
-    }
-
-    const triggerRows = this.db.prepare(`
-      SELECT trigger_type, COUNT(*) as count FROM executions GROUP BY trigger_type
-    `).all() as { trigger_type: string; count: number }[];
-
-    const by_trigger: Record<string, number> = {};
-    for (const r of triggerRows) {
-      by_trigger[r.trigger_type] = r.count;
-    }
-
-    return { total_executions: total, today_count: todayCount, by_skill, by_trigger, running };
-  }
-
-  /** Daily aggregated stats for the last N days */
-  dailyStats(days: number): {
-    date: string;
-    executions: number;
-    successes: number;
-    failures: number;
-    totalTokens: number;
-    inputTokens: number;
-    outputTokens: number;
-    cacheReadTokens: number;
-    costUsd: number;
-  }[] {
-    // Build the inclusive UTC date window: [today - (days-1), today].
-    // SQLite's date(started_at) returns a UTC YYYY-MM-DD string, so we
-    // generate the same format here to align keys.
-    const today = new Date();
-    const startUtc = new Date(Date.UTC(
-      today.getUTCFullYear(),
-      today.getUTCMonth(),
-      today.getUTCDate() - (days - 1),
-    ));
-
-    const dateKeys: string[] = [];
-    for (let i = 0; i < days; i++) {
-      const d = new Date(startUtc);
-      d.setUTCDate(startUtc.getUTCDate() + i);
-      dateKeys.push(d.toISOString().slice(0, 10));
-    }
-
-    const rows = this.db.prepare(`
-      SELECT
-        date(started_at) AS date,
-        COUNT(*) AS executions,
-        SUM(CASE WHEN success = 1 THEN 1 ELSE 0 END) AS successes,
-        SUM(CASE WHEN success = 0 THEN 1 ELSE 0 END) AS failures,
-        COALESCE(SUM(input_tokens), 0) + COALESCE(SUM(output_tokens), 0) + COALESCE(SUM(cache_read_input_tokens), 0) AS totalTokens,
-        COALESCE(SUM(input_tokens), 0) AS inputTokens,
-        COALESCE(SUM(output_tokens), 0) AS outputTokens,
-        COALESCE(SUM(cache_read_input_tokens), 0) AS cacheReadTokens,
-        COALESCE(SUM(cost_usd), 0) AS costUsd
-      FROM executions
-      WHERE date(started_at) >= ?
-      GROUP BY date(started_at)
-    `).all(dateKeys[0]) as {
-      date: string;
-      executions: number;
-      successes: number;
-      failures: number;
-      totalTokens: number;
-      inputTokens: number;
-      outputTokens: number;
-      cacheReadTokens: number;
-      costUsd: number;
-    }[];
-
-    const byDate = new Map(rows.map((r) => [r.date, r]));
-    return dateKeys.map((date) => byDate.get(date) ?? {
-      date,
-      executions: 0,
-      successes: 0,
-      failures: 0,
-      totalTokens: 0,
-      inputTokens: 0,
-      outputTokens: 0,
-      cacheReadTokens: 0,
-      costUsd: 0,
-    });
-  }
-
-  /** Hourly aggregated stats for the last N hours (UTC). Bucket key is `YYYY-MM-DDTHH`. */
-  hourlyStats(hours: number): {
-    date: string;
-    executions: number;
-    successes: number;
-    failures: number;
-    totalTokens: number;
-    inputTokens: number;
-    outputTokens: number;
-    cacheReadTokens: number;
-    costUsd: number;
-  }[] {
-    const now = new Date();
-    const startUtc = new Date(Date.UTC(
-      now.getUTCFullYear(),
-      now.getUTCMonth(),
-      now.getUTCDate(),
-      now.getUTCHours() - (hours - 1),
-    ));
-
-    const hourKeys: string[] = [];
-    for (let i = 0; i < hours; i++) {
-      const d = new Date(startUtc);
-      d.setUTCHours(startUtc.getUTCHours() + i);
-      // YYYY-MM-DDTHH (matches strftime('%Y-%m-%dT%H', …))
-      hourKeys.push(d.toISOString().slice(0, 13));
-    }
-
-    const rows = this.db.prepare(`
-      SELECT
-        strftime('%Y-%m-%dT%H', started_at) AS date,
-        COUNT(*) AS executions,
-        SUM(CASE WHEN success = 1 THEN 1 ELSE 0 END) AS successes,
-        SUM(CASE WHEN success = 0 THEN 1 ELSE 0 END) AS failures,
-        COALESCE(SUM(input_tokens), 0) + COALESCE(SUM(output_tokens), 0) + COALESCE(SUM(cache_read_input_tokens), 0) AS totalTokens,
-        COALESCE(SUM(input_tokens), 0) AS inputTokens,
-        COALESCE(SUM(output_tokens), 0) AS outputTokens,
-        COALESCE(SUM(cache_read_input_tokens), 0) AS cacheReadTokens,
-        COALESCE(SUM(cost_usd), 0) AS costUsd
-      FROM executions
-      WHERE strftime('%Y-%m-%dT%H', started_at) >= ?
-      GROUP BY strftime('%Y-%m-%dT%H', started_at)
-    `).all(hourKeys[0]) as {
-      date: string;
-      executions: number;
-      successes: number;
-      failures: number;
-      totalTokens: number;
-      inputTokens: number;
-      outputTokens: number;
-      cacheReadTokens: number;
-      costUsd: number;
-    }[];
-
-    const byHour = new Map(rows.map((r) => [r.date, r]));
-    return hourKeys.map((date) => byHour.get(date) ?? {
-      date,
-      executions: 0,
-      successes: 0,
-      failures: 0,
-      totalTokens: 0,
-      inputTokens: 0,
-      outputTokens: 0,
-      cacheReadTokens: 0,
-      costUsd: 0,
-    });
-  }
-
-  // ── Workflow Runs ──────────────────────────────────────────────
-
-  /** Create a new workflow run record */
-  createWorkflowRun(run: Omit<WorkflowRun, "phaseHistory" | "updatedAt">): void {
-    const now = new Date().toISOString();
-    this.db.prepare(`
-      INSERT INTO workflow_runs (id, workflow_name, trigger_id, repo, issue_number, current_phase, phase_history, status, context, scratch, started_at, updated_at)
-      VALUES (?, ?, ?, ?, ?, ?, '[]', ?, ?, ?, ?, ?)
-    `).run(
-      run.id,
-      run.workflowName,
-      run.triggerId,
-      run.repo ?? null,
-      run.issueNumber ?? null,
-      run.currentPhase,
-      run.status,
-      run.context ? JSON.stringify(run.context) : null,
-      run.scratch ? JSON.stringify(run.scratch) : null,
-      run.startedAt,
-      now,
-    );
-  }
-
-  /**
-   * Top-level merge of `patch` into the workflow run's scratch state.
-   * Loop iterations can append to `scratch.socratic.qa` without clobbering
-   * other keys.
-   */
-  updateWorkflowRunScratch(id: string, patch: Record<string, unknown>): void {
-    const now = new Date().toISOString();
-    const row = this.db
-      .prepare(`SELECT scratch FROM workflow_runs WHERE id = ?`)
-      .get(id) as { scratch: string | null } | undefined;
-    if (!row) return;
-    const current = row.scratch ? (JSON.parse(row.scratch) as Record<string, unknown>) : {};
-    const merged = { ...current, ...patch };
-    this.db
-      .prepare(`UPDATE workflow_runs SET scratch = ?, updated_at = ? WHERE id = ?`)
-      .run(JSON.stringify(merged), now, id);
-  }
-
-  /** Update the current phase and append to phase history */
-  updateWorkflowPhase(id: string, phase: string, entry: PhaseHistoryEntry): void {
-    const now = new Date().toISOString();
-    const row = this.db.prepare(`SELECT phase_history FROM workflow_runs WHERE id = ?`).get(id) as { phase_history: string } | undefined;
-    if (!row) return;
-    const history: PhaseHistoryEntry[] = JSON.parse(row.phase_history);
-    history.push(entry);
-    this.db.prepare(`
-      UPDATE workflow_runs SET current_phase = ?, phase_history = ?, updated_at = ? WHERE id = ?
-    `).run(phase, JSON.stringify(history), now, id);
-  }
-
-  /** Mark a workflow run as finished */
-  finishWorkflowRun(id: string, status: "succeeded" | "failed" | "cancelled", error?: string): void {
-    const now = new Date().toISOString();
-    this.db.prepare(`
-      UPDATE workflow_runs SET status = ?, finished_at = ?, updated_at = ?, context = CASE
-        WHEN ? IS NOT NULL THEN json_patch(COALESCE(context, '{}'), json_object('error', ?))
-        ELSE context
-      END WHERE id = ?
-    `).run(status, now, now, error ?? null, error ?? null, id);
-  }
-
-  /** Get a single workflow run by ID */
-  getWorkflowRun(id: string): WorkflowRun | null {
-    const row = this.db.prepare(`SELECT * FROM workflow_runs WHERE id = ?`).get(id) as Record<string, unknown> | undefined;
-    return row ? this.deserializeWorkflowRun(row) : null;
-  }
-
-  /** Find the most recent active (running or paused) workflow run for a trigger */
-  getWorkflowRunByTrigger(triggerId: string): WorkflowRun | null {
-    const row = this.db.prepare(`
-      SELECT * FROM workflow_runs
-      WHERE trigger_id = ? AND status IN ('running', 'paused')
-      ORDER BY started_at DESC
-      LIMIT 1
-    `).get(triggerId) as Record<string, unknown> | undefined;
-    return row ? this.deserializeWorkflowRun(row) : null;
-  }
-
-  /** List all active (running or paused) workflow runs */
-  activeWorkflowRuns(): WorkflowRun[] {
-    const rows = this.db.prepare(`
-      SELECT * FROM workflow_runs WHERE status IN ('running', 'paused') ORDER BY started_at DESC
-    `).all() as Record<string, unknown>[];
-    return rows.map((r) => this.deserializeWorkflowRun(r));
-  }
-
-  /** List recent workflow runs, ordered by start time descending */
-  recentWorkflowRuns(limit = 20): WorkflowRun[] {
-    const rows = this.db.prepare(`
-      SELECT * FROM workflow_runs ORDER BY started_at DESC LIMIT ?
-    `).all(limit) as Record<string, unknown>[];
-    return rows.map((r) => this.deserializeWorkflowRun(r));
-  }
-
-  /**
-   * List workflow runs with pagination + filters. Returns the page slice
-   * AND the post-filter total so the dashboard knows how many remain.
-   *
-   * - sinceIso filters `started_at >= sinceIso` (used for the header date range).
-   * - workflowName filters by exact `workflow_name`.
-   * - statuses filters to one of the workflow statuses (`running`, `paused`,
-   *   `succeeded`, `failed`, `cancelled`). Used for the dashboard's "live"
-   *   filter, which maps to ['running','paused'].
-   */
-  listWorkflowRuns(opts: {
-    limit?: number;
-    offset?: number;
-    sinceIso?: string;
-    workflowName?: string;
-    statuses?: string[];
-  } = {}): { runs: WorkflowRun[]; total: number } {
-    const limit = opts.limit ?? 20;
-    const offset = opts.offset ?? 0;
-
-    const where: string[] = [];
-    const params: unknown[] = [];
-    if (opts.sinceIso) {
-      where.push("started_at >= ?");
-      params.push(opts.sinceIso);
-    }
-    if (opts.workflowName) {
-      where.push("workflow_name = ?");
-      params.push(opts.workflowName);
-    }
-    if (opts.statuses && opts.statuses.length > 0) {
-      where.push(`status IN (${opts.statuses.map(() => "?").join(",")})`);
-      params.push(...opts.statuses);
-    }
-    const whereClause = where.length > 0 ? `WHERE ${where.join(" AND ")}` : "";
-
-    const total = (
-      this.db
-        .prepare(`SELECT COUNT(*) as c FROM workflow_runs ${whereClause}`)
-        .get(...params) as { c: number }
-    ).c;
-
-    // Explicit column list — the heavy JSON blobs (`context`, `scratch`)
-    // can each be many MB on long-running build runs and
-    // the dashboard list view doesn't read them. Returning them turned a
-    // 20-row page into a 14MB payload. The single-row endpoint
-    // (`getWorkflowRun`) still uses `SELECT *` so the detail panel keeps
-    // the full row when the user picks one.
-    const rows = this.db
-      .prepare(
-        `SELECT
-           id, workflow_name, trigger_id, repo, issue_number,
-           current_phase, phase_history, status,
-           restart_count, started_at, updated_at, finished_at
-         FROM workflow_runs
-         ${whereClause}
-         ORDER BY started_at DESC
-         LIMIT ? OFFSET ?`,
-      )
-      .all(...params, limit, offset) as Record<string, unknown>[];
-
-    return {
-      runs: rows.map((r) => this.deserializeWorkflowRun(r)),
-      total,
-    };
-  }
-
-  /** Distinct workflow_name values, sorted alphabetically. */
-  distinctWorkflowNames(): string[] {
-    const rows = this.db
-      .prepare(`SELECT DISTINCT workflow_name FROM workflow_runs ORDER BY workflow_name ASC`)
-      .all() as { workflow_name: string }[];
-    return rows.map((r) => r.workflow_name);
-  }
-
-  /** Cancel a workflow run */
-  cancelWorkflowRun(id: string): void {
-    const now = new Date().toISOString();
-    this.db.prepare(`
-      UPDATE workflow_runs SET status = 'cancelled', updated_at = ?, finished_at = ? WHERE id = ?
-    `).run(now, now, id);
-  }
-
-  /** Pause a workflow run (waiting for approval) */
-  pauseWorkflowRun(id: string): void {
-    const now = new Date().toISOString();
-    this.db.prepare(`
-      UPDATE workflow_runs SET status = 'paused', updated_at = ? WHERE id = ?
-    `).run(now, id);
-  }
-
-  /** Resume a paused workflow run (set back to running) */
-  resumeWorkflowRun(id: string): void {
-    const now = new Date().toISOString();
-    this.db.prepare(`
-      UPDATE workflow_runs SET status = 'running', updated_at = ? WHERE id = ?
-    `).run(now, id);
-  }
-
-  /**
-   * Increment the restart counter and return the new value. Used by
-   * `resumeOrphanedWorkflows` to enforce a retry budget so a run that
-   * crashes the host (agent OOM, etc.) eventually self-terminates instead
-   * of being re-dispatched on every boot.
-   */
-  incrementWorkflowRunRestartCount(id: string): number {
-    const now = new Date().toISOString();
-    this.db.prepare(`
-      UPDATE workflow_runs
-      SET restart_count = COALESCE(restart_count, 0) + 1, updated_at = ?
-      WHERE id = ?
-    `).run(now, id);
-    const row = this.db.prepare(`SELECT restart_count FROM workflow_runs WHERE id = ?`)
-      .get(id) as { restart_count: number } | undefined;
-    return row?.restart_count ?? 0;
-  }
-
-  // ── Workflow Approvals ─────────────────────────────────────────
-
-  /** Create a new pending approval request */
-  createApproval(
-    approval: Omit<WorkflowApproval, "status" | "respondedBy" | "response" | "respondedAt" | "kind"> & {
-      kind?: WorkflowApproval["kind"];
-    },
-  ): void {
-    this.db
-      .prepare(
-        `
-      INSERT INTO workflow_approvals (id, workflow_run_id, gate, summary, status, kind, requested_by, created_at)
-      VALUES (?, ?, ?, ?, 'pending', ?, ?, ?)
-    `,
-      )
-      .run(
-        approval.id,
-        approval.workflowRunId,
-        approval.gate,
-        approval.summary,
-        approval.kind ?? "approve",
-        approval.requestedBy ?? null,
-        approval.createdAt,
-      );
-  }
-
-  /**
-   * Resolve a reply gate: marks the approval row as approved (reply gates
-   * don't have approve/reject semantics — any reply is a "go") and stores
-   * the reply text as the `response`. Used by the socratic explore loop.
-   */
-  resolveReplyGate(id: string, replyText: string, responder: string): void {
-    const now = new Date().toISOString();
-    this.db
-      .prepare(
-        `UPDATE workflow_approvals
-           SET status = 'approved',
-               responded_by = ?,
-               response = ?,
-               responded_at = ?
-         WHERE id = ? AND kind = 'reply'`,
-      )
-      .run(responder, replyText, now, id);
-  }
-
-  /**
-   * Find the most recent pending reply gate for a given trigger id. Used by
-   * the router to short-circuit free-form replies on a paused socratic
-   * explore loop without re-running classifier logic.
-   */
-  getPendingReplyGateByTrigger(triggerId: string): WorkflowApproval | null {
-    const row = this.db
-      .prepare(
-        `SELECT wa.* FROM workflow_approvals wa
-         JOIN workflow_runs wr ON wa.workflow_run_id = wr.id
-         WHERE wr.trigger_id = ? AND wa.status = 'pending' AND wa.kind = 'reply'
-         ORDER BY wa.created_at DESC
-         LIMIT 1`,
-      )
-      .get(triggerId) as Record<string, unknown> | undefined;
-    return row ? this.deserializeApproval(row) : null;
-  }
-
-  /** Get a single approval by ID */
-  getApproval(id: string): WorkflowApproval | null {
-    const row = this.db.prepare(`SELECT * FROM workflow_approvals WHERE id = ?`).get(id) as Record<string, unknown> | undefined;
-    return row ? this.deserializeApproval(row) : null;
-  }
-
-  /** Find the pending approval for a workflow run */
-  getPendingApprovalForWorkflow(workflowRunId: string): WorkflowApproval | null {
-    const row = this.db.prepare(`
-      SELECT * FROM workflow_approvals WHERE workflow_run_id = ? AND status = 'pending' LIMIT 1
-    `).get(workflowRunId) as Record<string, unknown> | undefined;
-    return row ? this.deserializeApproval(row) : null;
-  }
-
-  /** Find the pending approval by trigger ID (join with workflow_runs) */
-  getPendingApprovalByTrigger(triggerId: string): WorkflowApproval | null {
-    const row = this.db.prepare(`
-      SELECT wa.* FROM workflow_approvals wa
-      JOIN workflow_runs wr ON wa.workflow_run_id = wr.id
-      WHERE wr.trigger_id = ? AND wa.status = 'pending'
-      ORDER BY wa.created_at DESC
-      LIMIT 1
-    `).get(triggerId) as Record<string, unknown> | undefined;
-    return row ? this.deserializeApproval(row) : null;
-  }
-
-  /** List all pending approvals */
-  listPendingApprovals(): WorkflowApproval[] {
-    const rows = this.db.prepare(`
-      SELECT * FROM workflow_approvals WHERE status = 'pending' ORDER BY created_at DESC
-    `).all() as Record<string, unknown>[];
-    return rows.map((r) => this.deserializeApproval(r));
-  }
-
-  /** Record the response to an approval */
-  respondToApproval(id: string, status: 'approved' | 'rejected', respondedBy: string, response?: string): void {
-    const now = new Date().toISOString();
-    this.db.prepare(`
-      UPDATE workflow_approvals SET status = ?, responded_by = ?, response = ?, responded_at = ? WHERE id = ?
-    `).run(status, respondedBy, response ?? null, now, id);
+    migrate(this.db);
+
+    this.executions = new ExecutionStore(this.db);
+    this.approvals = new ApprovalStore(this.db);
+    this.runs = new WorkflowRunStore(this.db, { approvals: this.approvals });
   }
 
   // ── Cron overrides ─────────────────────────────────────────────
@@ -1379,41 +196,6 @@ export class StateDb {
       enabled: (row.enabled as number) === 1,
       updatedAt: row.updated_at as string,
       updatedBy: (row.updated_by as string | null) ?? null,
-    };
-  }
-
-  private deserializeApproval(row: Record<string, unknown>): WorkflowApproval {
-    return {
-      id: row.id as string,
-      workflowRunId: row.workflow_run_id as string,
-      gate: row.gate as string,
-      summary: row.summary as string,
-      status: row.status as WorkflowApproval['status'],
-      kind: ((row.kind as string | undefined) || "approve") as WorkflowApproval["kind"],
-      requestedBy: row.requested_by as string | undefined || undefined,
-      respondedBy: row.responded_by as string | undefined || undefined,
-      response: row.response as string | undefined || undefined,
-      respondedAt: row.responded_at as string | undefined || undefined,
-      createdAt: row.created_at as string,
-    };
-  }
-
-  private deserializeWorkflowRun(row: Record<string, unknown>): WorkflowRun {
-    return {
-      id: row.id as string,
-      workflowName: row.workflow_name as string,
-      triggerId: row.trigger_id as string,
-      repo: row.repo as string | undefined,
-      issueNumber: row.issue_number as number | undefined,
-      currentPhase: row.current_phase as string,
-      phaseHistory: JSON.parse(row.phase_history as string) as PhaseHistoryEntry[],
-      status: row.status as WorkflowRun["status"],
-      context: row.context ? JSON.parse(row.context as string) as Record<string, unknown> : undefined,
-      scratch: row.scratch ? JSON.parse(row.scratch as string) as Record<string, unknown> : undefined,
-      restartCount: typeof row.restart_count === "number" ? row.restart_count : 0,
-      startedAt: row.started_at as string,
-      updatedAt: row.updated_at as string,
-      finishedAt: row.finished_at as string | undefined,
     };
   }
