+++ src/state/execution-store.ts
@@ -1,0 +1,715 @@
+import type Database from "better-sqlite3";
+import { randomUUID } from "crypto";
+
+export interface ExecutionRecord {
+  id: string;
+  triggerType: "webhook" | "cron" | "chat" | "api";
+  triggerId: string;
+  skill: string;
+  repo?: string;
+  issueNumber?: number;
+  startedAt: string;
+  finishedAt?: string;
+  success?: boolean;
+  error?: string;
+  turns?: number;
+  durationMs?: number;
+  /**
+   * Runtime session id (OpenCode `ses_…`) captured from the event stream.
+   * Used by the dashboard to look up the session log for a given phase.
+   */
+  sessionId?: string;
+  /** Total USD cost reported by the runtime on the final step. Zero under OAuth/subscription auth. */
+  costUsd?: number;
+  /** Fresh input tokens (no cache hit). */
+  inputTokens?: number;
+  /** Tokens that triggered a cache write. */
+  cacheCreationInputTokens?: number;
+  /** Tokens served from prompt cache. */
+  cacheReadInputTokens?: number;
+  /** Output tokens generated by the model. */
+  outputTokens?: number;
+  /** API-side duration (excludes orchestrator overhead). */
+  apiDurationMs?: number;
+  /** Result subtype, e.g. "success" or "error_max_turns". */
+  stopReason?: string;
+  /**
+   * JSON-serialized {@link ExtensionStatusMap} — which agentic-pi extensions
+   * (file-search / github / web-search) were active for this execution.
+   */
+  extensionStatus?: string;
+  /**
+   * Workflow run that owns this execution. Scopes dedup: a fresh re-trigger
+   * creates a new workflow_run_id, so `shouldRunPhase` won't find matching
+   * rows from the prior run and will correctly run the phase again.
+   */
+  workflowRunId?: string;
+}
+
+/**
+ * Owns every read/write against the `executions` table — the harness's
+ * append-only ledger of agent runs. The dashboard derives phase status from
+ * this table (primary source, ahead of the `phase_history` fallback), and the
+ * runner's resume path skips already-completed phases via {@link shouldRunPhase}.
+ *
+ * Carved out of the old `StateDb` god-class (issue #97). Constructed from the
+ * single shared `Database` so it sits in the same connection / transaction
+ * scope as the other stores.
+ */
+export class ExecutionStore {
+  constructor(private db: Database.Database) {}
+
+  recordStart(record: Omit<ExecutionRecord, "finishedAt" | "success" | "error" | "turns" | "durationMs">): void {
+    this.db.prepare(`
+      INSERT INTO executions (id, trigger_type, trigger_id, skill, repo, issue_number, started_at, workflow_run_id)
+      VALUES (?, ?, ?, ?, ?, ?, ?, ?)
+    `).run(
+      record.id,
+      record.triggerType,
+      record.triggerId,
+      record.skill,
+      record.repo,
+      record.issueNumber,
+      record.startedAt,
+      record.workflowRunId ?? null,
+    );
+  }
+
+  /**
+   * Mid-run update: persist the agent session id as soon as the executor sees
+   * the stream-json `system/init` line. Lets the dashboard surface the live
+   * session log for an in-flight phase, not just for completed ones.
+   */
+  recordSessionId(id: string, sessionId: string): void {
+    this.db.prepare(`UPDATE executions SET session_id = ? WHERE id = ?`).run(sessionId, id);
+  }
+
+  /**
+   * Persist the final assistant text for a loop-iteration execution.
+   * Used by the runner when a scratch slot would otherwise inline the
+   * full LLM output into `workflow_runs.scratch` — the scratch entry
+   * stores `lastOutputExecutionId` and resolves the text through
+   * `getExecutionOutput` on read.
+   */
+  recordOutputText(id: string, output: string): void {
+    this.db.prepare(`UPDATE executions SET output_text = ? WHERE id = ?`).run(output, id);
+  }
+
+  /** Read the persisted output text for an execution, or null if none. */
+  getExecutionOutput(id: string): string | null {
+    const row = this.db
+      .prepare(`SELECT output_text FROM executions WHERE id = ?`)
+      .get(id) as { output_text: string | null } | undefined;
+    return row?.output_text ?? null;
+  }
+
+  /**
+   * Most recent finished execution's `output_text` for a phase (`skill`), or
+   * null. Used by the reviewer loop on resume to re-derive a completed review's
+   * verdict from its persisted output (a dedup-`done` review must not be
+   * assumed APPROVED — it may have requested changes).
+   */
+  getPhaseOutput(skill: string, triggerId: string, workflowRunId?: string): string | null {
+    const [scopeClause, scopeParam] = workflowRunId
+      ? ["workflow_run_id = ?", workflowRunId]
+      : ["trigger_id = ?", triggerId];
+    const row = this.db.prepare(`
+      SELECT output_text FROM executions
+      WHERE skill = ? AND ${scopeClause} AND output_text IS NOT NULL
+      ORDER BY finished_at DESC
+      LIMIT 1
+    `).get(skill, scopeParam) as { output_text: string | null } | undefined;
+    return row?.output_text ?? null;
+  }
+
+  recordFinish(
+    id: string,
+    result: {
+      success: boolean;
+      error?: string;
+      turns?: number;
+      durationMs?: number;
+      sessionId?: string;
+      costUsd?: number;
+      inputTokens?: number;
+      cacheCreationInputTokens?: number;
+      cacheReadInputTokens?: number;
+      outputTokens?: number;
+      apiDurationMs?: number;
+      stopReason?: string;
+      /** JSON-serialized ExtensionStatusMap (extensions active this run). */
+      extensionStatus?: string;
+    },
+  ): void {
+    this.db.prepare(`
+      UPDATE executions
+      SET finished_at = ?,
+          success = ?,
+          error = ?,
+          turns = ?,
+          duration_ms = ?,
+          session_id = COALESCE(?, session_id),
+          cost_usd = COALESCE(?, cost_usd),
+          input_tokens = COALESCE(?, input_tokens),
+          cache_creation_input_tokens = COALESCE(?, cache_creation_input_tokens),
+          cache_read_input_tokens = COALESCE(?, cache_read_input_tokens),
+          output_tokens = COALESCE(?, output_tokens),
+          api_duration_ms = COALESCE(?, api_duration_ms),
+          stop_reason = COALESCE(?, stop_reason),
+          extension_status = COALESCE(?, extension_status)
+      WHERE id = ?
+    `).run(
+      new Date().toISOString(),
+      result.success ? 1 : 0,
+      result.error,
+      result.turns,
+      result.durationMs,
+      result.sessionId ?? null,
+      result.costUsd ?? null,
+      result.inputTokens ?? null,
+      result.cacheCreationInputTokens ?? null,
+      result.cacheReadInputTokens ?? null,
+      result.outputTokens ?? null,
+      result.apiDurationMs ?? null,
+      result.stopReason ?? null,
+      result.extensionStatus ?? null,
+      id,
+    );
+  }
+
+  /**
+   * Record a phase that the scheduler skipped because its trigger rule was
+   * not satisfied (a failed/skipped upstream cascaded down a chain). The
+   * executions ledger is the single source of truth the dashboard derives
+   * phase status from, so skips are written here too — as a finished,
+   * non-successful row (`success = 0`, `stop_reason = 'skipped'`). Because it
+   * is not `success = 1`, `shouldRunPhase` re-evaluates it on resume (the node
+   * will simply be re-skipped if the upstream is still failed).
+   */
+  recordSkippedPhase(
+    skill: string,
+    triggerId: string,
+    workflowRunId?: string,
+    repo?: string,
+  ): void {
+    const now = new Date().toISOString();
+    const m = triggerId.match(/#(\d+)$/);
+    const issueNumber = m ? Number(m[1]) : null;
+    this.db.prepare(`
+      INSERT INTO executions
+        (id, trigger_type, trigger_id, skill, repo, issue_number, started_at, finished_at, success, error, stop_reason, workflow_run_id)
+      VALUES (?, 'webhook', ?, ?, ?, ?, ?, ?, 0, 'skipped: trigger rule not satisfied', 'skipped', ?)
+    `).run(
+      randomUUID(),
+      triggerId,
+      skill,
+      repo ?? null,
+      issueNumber,
+      now,
+      now,
+      workflowRunId ?? null,
+    );
+  }
+
+  /**
+   * List chat threads aggregated from the executions table.
+   *
+   * Each row collapses every execution sharing a `trigger_id` (the messaging
+   * session id) into one summary so the dashboard's chat-sessions tab can
+   * show one card per Slack thread, not one per turn. Joined with
+   * `messaging_sessions` for the agent SDK session id (used to resolve to
+   * the on-disk jsonl when streaming messages) and with `messaging_messages`
+   * for the most recent assistant text (preview snippet).
+   */
+  listChatThreads(limit: number): {
+    triggerId: string;
+    agentSessionId: string | null;
+    firstStartedAt: string;
+    lastActivityAt: string;
+    turnCount: number;
+    totalCost: number;
+    inputTokens: number;
+    outputTokens: number;
+    cacheReadTokens: number;
+    lastAssistantContent: string | null;
+  }[] {
+    const rows = this.db.prepare(`
+      SELECT
+        e.trigger_id              AS triggerId,
+        ms.agent_session_id       AS agentSessionId,
+        MIN(e.started_at)         AS firstStartedAt,
+        MAX(COALESCE(e.finished_at, e.started_at)) AS lastActivityAt,
+        COUNT(*)                  AS turnCount,
+        SUM(COALESCE(e.cost_usd, 0))           AS totalCost,
+        SUM(COALESCE(e.input_tokens, 0))       AS inputTokens,
+        SUM(COALESCE(e.output_tokens, 0))      AS outputTokens,
+        SUM(COALESCE(e.cache_read_input_tokens, 0)) AS cacheReadTokens,
+        (
+          SELECT mm.content
+          FROM messaging_messages mm
+          WHERE mm.session_id = e.trigger_id AND mm.role = 'assistant'
+          ORDER BY mm.timestamp DESC
+          LIMIT 1
+        ) AS lastAssistantContent
+      FROM executions e
+      LEFT JOIN messaging_sessions ms ON ms.id = e.trigger_id
+      WHERE e.skill = 'chat'
+      GROUP BY e.trigger_id
+      ORDER BY lastActivityAt DESC
+      LIMIT ?
+    `).all(limit) as Array<{
+      triggerId: string;
+      agentSessionId: string | null;
+      firstStartedAt: string;
+      lastActivityAt: string;
+      turnCount: number;
+      totalCost: number;
+      inputTokens: number;
+      outputTokens: number;
+      cacheReadTokens: number;
+      lastAssistantContent: string | null;
+    }>;
+    return rows;
+  }
+
+  /** Look up a single chat thread by its messaging session id (= trigger_id). */
+  getChatThread(triggerId: string): {
+    triggerId: string;
+    agentSessionId: string | null;
+    firstStartedAt: string;
+    lastActivityAt: string;
+    turnCount: number;
+    totalCost: number;
+    inputTokens: number;
+    outputTokens: number;
+    cacheReadTokens: number;
+    lastAssistantContent: string | null;
+  } | null {
+    const row = this.db.prepare(`
+      SELECT
+        e.trigger_id              AS triggerId,
+        ms.agent_session_id       AS agentSessionId,
+        MIN(e.started_at)         AS firstStartedAt,
+        MAX(COALESCE(e.finished_at, e.started_at)) AS lastActivityAt,
+        COUNT(*)                  AS turnCount,
+        SUM(COALESCE(e.cost_usd, 0))           AS totalCost,
+        SUM(COALESCE(e.input_tokens, 0))       AS inputTokens,
+        SUM(COALESCE(e.output_tokens, 0))      AS outputTokens,
+        SUM(COALESCE(e.cache_read_input_tokens, 0)) AS cacheReadTokens,
+        (
+          SELECT mm.content
+          FROM messaging_messages mm
+          WHERE mm.session_id = e.trigger_id AND mm.role = 'assistant'
+          ORDER BY mm.timestamp DESC
+          LIMIT 1
+        ) AS lastAssistantContent
+      FROM executions e
+      LEFT JOIN messaging_sessions ms ON ms.id = e.trigger_id
+      WHERE e.skill = 'chat' AND e.trigger_id = ?
+      GROUP BY e.trigger_id
+    `).get(triggerId) as {
+      triggerId: string;
+      agentSessionId: string | null;
+      firstStartedAt: string;
+      lastActivityAt: string;
+      turnCount: number;
+      totalCost: number;
+      inputTokens: number;
+      outputTokens: number;
+      cacheReadTokens: number;
+      lastAssistantContent: string | null;
+    } | undefined;
+    return row ?? null;
+  }
+
+  /** Check if a skill is currently running for a given trigger */
+  isRunning(skill: string, triggerId: string): boolean {
+    const row = this.db.prepare(`
+      SELECT 1 FROM executions
+      WHERE skill = ? AND trigger_id = ? AND finished_at IS NULL
+      LIMIT 1
+    `).get(skill, triggerId);
+    return !!row;
+  }
+
+  /** Check if a skill has already completed successfully for a given trigger */
+  isCompleted(skill: string, triggerId: string): boolean {
+    const row = this.db.prepare(`
+      SELECT 1 FROM executions
+      WHERE skill = ? AND trigger_id = ? AND success = 1
+      LIMIT 1
+    `).get(skill, triggerId);
+    return !!row;
+  }
+
+  /**
+   * Check if a phase should run: not currently running AND not already
+   * succeeded. When `workflowRunId` is provided, scopes the check to that
+   * specific run — a fresh re-trigger (new workflow_run_id) will see no
+   * matching rows and correctly return "run" even if a prior completed run
+   * for the same trigger exists. Omit `workflowRunId` for legacy callers
+   * (e.g. chat/status-report dedup) that want the old trigger-wide scope.
+   */
+  shouldRunPhase(skill: string, triggerId: string, workflowRunId?: string): "run" | "running" | "done" {
+    const [scopeClause, scopeParam] = workflowRunId
+      ? ["workflow_run_id = ?", workflowRunId]
+      : ["trigger_id = ?", triggerId];
+
+    const running = this.db.prepare(`
+      SELECT 1 FROM executions
+      WHERE skill = ? AND ${scopeClause} AND finished_at IS NULL
+      LIMIT 1
+    `).get(skill, scopeParam);
+    if (running) return "running";
+
+    const done = this.db.prepare(`
+      SELECT 1 FROM executions
+      WHERE skill = ? AND ${scopeClause} AND success = 1
+      LIMIT 1
+    `).get(skill, scopeParam);
+    if (done) return "done";
+
+    return "run";
+  }
+
+  /** Mark all stale "running" executions for a skill/trigger as failed.
+   *  Called when we detect no matching Docker container is alive. */
+  markStaleAsFailed(skill: string, triggerId: string, workflowRunId?: string): number {
+    const [scopeClause, scopeParam] = workflowRunId
+      ? ["workflow_run_id = ?", workflowRunId]
+      : ["trigger_id = ?", triggerId];
+    const result = this.db.prepare(`
+      UPDATE executions
+      SET finished_at = ?, success = 0, error = 'stale: container no longer running'
+      WHERE skill = ? AND ${scopeClause} AND finished_at IS NULL
+    `).run(new Date().toISOString(), skill, scopeParam);
+    return result.changes;
+  }
+
+  /**
+   * Mark every unfinished execution row for a trigger as failed, regardless
+   * of skill. Used by the boot-time recovery sweep to clear out phases that
+   * were running when the harness crashed/restarted, so the runner's dedup
+   * logic doesn't think they're still in flight.
+   */
+  markAllStaleForTrigger(triggerId: string, reason: string): number {
+    const result = this.db.prepare(`
+      UPDATE executions
+      SET finished_at = ?, success = 0, error = ?
+      WHERE trigger_id = ? AND finished_at IS NULL
+    `).run(new Date().toISOString(), reason, triggerId);
+    return result.changes;
+  }
+
+  /**
+   * Mark the most recent execution for a skill/trigger as failed.
+   * Used when the agent SDK reported success but the orchestrator decided
+   * the business outcome was a failure (e.g. guardrails BLOCKED, reviewer
+   * REQUEST_CHANGES). Without this, runPhase would skip the phase on the
+   * next run because the DB still says success=1.
+   */
+  markLatestAsFailed(skill: string, triggerId: string, reason: string, workflowRunId?: string): number {
+    const [scopeClause, scopeParam] = workflowRunId
+      ? ["workflow_run_id = ?", workflowRunId]
+      : ["trigger_id = ?", triggerId];
+    const result = this.db.prepare(`
+      UPDATE executions
+      SET success = 0, error = ?
+      WHERE id = (
+        SELECT id FROM executions
+        WHERE skill = ? AND ${scopeClause}
+        ORDER BY started_at DESC
+        LIMIT 1
+      )
+    `).run(reason, skill, scopeParam);
+    return result.changes;
+  }
+
+  /** Get recent executions for a skill */
+  recentExecutions(skill: string, limit = 10): ExecutionRecord[] {
+    return this.db.prepare(`
+      SELECT * FROM executions
+      WHERE skill = ?
+      ORDER BY started_at DESC
+      LIMIT ?
+    `).all(skill, limit) as ExecutionRecord[];
+  }
+
+  /** Count consecutive failures for a skill (for cron failure tracking) */
+  consecutiveFailures(skill: string): number {
+    const rows = this.db.prepare(`
+      SELECT success FROM executions
+      WHERE skill = ? AND finished_at IS NOT NULL
+      ORDER BY started_at DESC
+      LIMIT 10
+    `).all(skill) as { success: number }[];
+
+    let count = 0;
+    for (const row of rows) {
+      if (row.success === 0) count++;
+      else break;
+    }
+    return count;
+  }
+
+  /** Get all executions with pagination */
+  allExecutions(limit = 100, offset = 0): ExecutionRecord[] {
+    return this.db.prepare(`
+      SELECT * FROM executions
+      ORDER BY started_at DESC
+      LIMIT ? OFFSET ?
+    `).all(limit, offset) as ExecutionRecord[];
+  }
+
+  /**
+   * Get every execution recorded for a workflow run, with camelCase fields
+   * ready for the dashboard. Ordered by start time ascending so phases
+   * display in the order they ran. Scopes to a specific `workflow_run_id`
+   * when provided (preferred — isolates fresh re-triggers from prior runs);
+   * falls back to the legacy trigger-wide scope for rows written before the
+   * workflow_run_id column existed.
+   */
+  getExecutionsForWorkflowRun(workflowRunId: string, triggerId: string, workflowName?: string): ExecutionRecord[] {
+    const skillPattern = workflowName ? `${workflowName}:%` : "%:%";
+    const rows = this.db.prepare(`
+      SELECT
+        id,
+        trigger_type    AS triggerType,
+        trigger_id      AS triggerId,
+        skill,
+        repo,
+        issue_number    AS issueNumber,
+        started_at      AS startedAt,
+        finished_at     AS finishedAt,
+        success,
+        error,
+        turns,
+        duration_ms     AS durationMs,
+        session_id      AS sessionId,
+        cost_usd        AS costUsd,
+        input_tokens    AS inputTokens,
+        cache_creation_input_tokens AS cacheCreationInputTokens,
+        cache_read_input_tokens     AS cacheReadInputTokens,
+        output_tokens   AS outputTokens,
+        api_duration_ms AS apiDurationMs,
+        stop_reason     AS stopReason,
+        extension_status AS extensionStatus,
+        workflow_run_id AS workflowRunId
+      FROM executions
+      WHERE (workflow_run_id = ? OR (workflow_run_id IS NULL AND trigger_id = ?))
+        AND skill LIKE ?
+      ORDER BY started_at ASC
+    `).all(workflowRunId, triggerId, skillPattern) as Array<Record<string, unknown>>;
+
+    return rows.map((r) => ({
+      id: r.id as string,
+      triggerType: r.triggerType as ExecutionRecord["triggerType"],
+      triggerId: r.triggerId as string,
+      skill: r.skill as string,
+      repo: (r.repo as string | null) ?? undefined,
+      issueNumber: (r.issueNumber as number | null) ?? undefined,
+      startedAt: r.startedAt as string,
+      finishedAt: (r.finishedAt as string | null) ?? undefined,
+      success: r.success === null || r.success === undefined ? undefined : Boolean(r.success),
+      error: (r.error as string | null) ?? undefined,
+      turns: (r.turns as number | null) ?? undefined,
+      durationMs: (r.durationMs as number | null) ?? undefined,
+      sessionId: (r.sessionId as string | null) ?? undefined,
+      costUsd: (r.costUsd as number | null) ?? undefined,
+      inputTokens: (r.inputTokens as number | null) ?? undefined,
+      cacheCreationInputTokens: (r.cacheCreationInputTokens as number | null) ?? undefined,
+      cacheReadInputTokens: (r.cacheReadInputTokens as number | null) ?? undefined,
+      outputTokens: (r.outputTokens as number | null) ?? undefined,
+      apiDurationMs: (r.apiDurationMs as number | null) ?? undefined,
+      stopReason: (r.stopReason as string | null) ?? undefined,
+      extensionStatus: (r.extensionStatus as string | null) ?? undefined,
+      workflowRunId: (r.workflowRunId as string | null) ?? undefined,
+    }));
+  }
+
+  /** Get currently running executions (no finished_at) */
+  runningExecutions(): ExecutionRecord[] {
+    return this.db.prepare(`
+      SELECT * FROM executions
+      WHERE finished_at IS NULL
+      ORDER BY started_at DESC
+    `).all() as ExecutionRecord[];
+  }
+
+  /** Aggregate execution stats */
+  executionStats(): {
+    total_executions: number;
+    today_count: number;
+    by_skill: Record<string, { count: number; success: number; fail: number }>;
+    by_trigger: Record<string, number>;
+    running: number;
+  } {
+    const today = new Date();
+    today.setHours(0, 0, 0, 0);
+    const todayIso = today.toISOString();
+
+    const total = (this.db.prepare(`SELECT COUNT(*) as c FROM executions`).get() as { c: number }).c;
+    const todayCount = (this.db.prepare(`SELECT COUNT(*) as c FROM executions WHERE started_at >= ?`).get(todayIso) as { c: number }).c;
+    const running = (this.db.prepare(`SELECT COUNT(*) as c FROM executions WHERE finished_at IS NULL`).get() as { c: number }).c;
+
+    const skillRows = this.db.prepare(`
+      SELECT skill, COUNT(*) as count,
+        SUM(CASE WHEN success = 1 THEN 1 ELSE 0 END) as success,
+        SUM(CASE WHEN success = 0 THEN 1 ELSE 0 END) as fail
+      FROM executions GROUP BY skill
+    `).all() as { skill: string; count: number; success: number; fail: number }[];
+
+    const by_skill: Record<string, { count: number; success: number; fail: number }> = {};
+    for (const r of skillRows) {
+      by_skill[r.skill] = { count: r.count, success: r.success, fail: r.fail };
+    }
+
+    const triggerRows = this.db.prepare(`
+      SELECT trigger_type, COUNT(*) as count FROM executions GROUP BY trigger_type
+    `).all() as { trigger_type: string; count: number }[];
+
+    const by_trigger: Record<string, number> = {};
+    for (const r of triggerRows) {
+      by_trigger[r.trigger_type] = r.count;
+    }
+
+    return { total_executions: total, today_count: todayCount, by_skill, by_trigger, running };
+  }
+
+  /** Daily aggregated stats for the last N days */
+  dailyStats(days: number): {
+    date: string;
+    executions: number;
+    successes: number;
+    failures: number;
+    totalTokens: number;
+    inputTokens: number;
+    outputTokens: number;
+    cacheReadTokens: number;
+    costUsd: number;
+  }[] {
+    // Build the inclusive UTC date window: [today - (days-1), today].
+    // SQLite's date(started_at) returns a UTC YYYY-MM-DD string, so we
+    // generate the same format here to align keys.
+    const today = new Date();
+    const startUtc = new Date(Date.UTC(
+      today.getUTCFullYear(),
+      today.getUTCMonth(),
+      today.getUTCDate() - (days - 1),
+    ));
+
+    const dateKeys: string[] = [];
+    for (let i = 0; i < days; i++) {
+      const d = new Date(startUtc);
+      d.setUTCDate(startUtc.getUTCDate() + i);
+      dateKeys.push(d.toISOString().slice(0, 10));
+    }
+
+    const rows = this.db.prepare(`
+      SELECT
+        date(started_at) AS date,
+        COUNT(*) AS executions,
+        SUM(CASE WHEN success = 1 THEN 1 ELSE 0 END) AS successes,
+        SUM(CASE WHEN success = 0 THEN 1 ELSE 0 END) AS failures,
+        COALESCE(SUM(input_tokens), 0) + COALESCE(SUM(output_tokens), 0) + COALESCE(SUM(cache_read_input_tokens), 0) AS totalTokens,
+        COALESCE(SUM(input_tokens), 0) AS inputTokens,
+        COALESCE(SUM(output_tokens), 0) AS outputTokens,
+        COALESCE(SUM(cache_read_input_tokens), 0) AS cacheReadTokens,
+        COALESCE(SUM(cost_usd), 0) AS costUsd
+      FROM executions
+      WHERE date(started_at) >= ?
+      GROUP BY date(started_at)
+    `).all(dateKeys[0]) as {
+      date: string;
+      executions: number;
+      successes: number;
+      failures: number;
+      totalTokens: number;
+      inputTokens: number;
+      outputTokens: number;
+      cacheReadTokens: number;
+      costUsd: number;
+    }[];
+
+    const byDate = new Map(rows.map((r) => [r.date, r]));
+    return dateKeys.map((date) => byDate.get(date) ?? {
+      date,
+      executions: 0,
+      successes: 0,
+      failures: 0,
+      totalTokens: 0,
+      inputTokens: 0,
+      outputTokens: 0,
+      cacheReadTokens: 0,
+      costUsd: 0,
+    });
+  }
+
+  /** Hourly aggregated stats for the last N hours (UTC). Bucket key is `YYYY-MM-DDTHH`. */
+  hourlyStats(hours: number): {
+    date: string;
+    executions: number;
+    successes: number;
+    failures: number;
+    totalTokens: number;
+    inputTokens: number;
+    outputTokens: number;
+    cacheReadTokens: number;
+    costUsd: number;
+  }[] {
+    const now = new Date();
+    const startUtc = new Date(Date.UTC(
+      now.getUTCFullYear(),
+      now.getUTCMonth(),
+      now.getUTCDate(),
+      now.getUTCHours() - (hours - 1),
+    ));
+
+    const hourKeys: string[] = [];
+    for (let i = 0; i < hours; i++) {
+      const d = new Date(startUtc);
+      d.setUTCHours(startUtc.getUTCHours() + i);
+      // YYYY-MM-DDTHH (matches strftime('%Y-%m-%dT%H', …))
+      hourKeys.push(d.toISOString().slice(0, 13));
+    }
+
+    const rows = this.db.prepare(`
+      SELECT
+        strftime('%Y-%m-%dT%H', started_at) AS date,
+        COUNT(*) AS executions,
+        SUM(CASE WHEN success = 1 THEN 1 ELSE 0 END) AS successes,
+        SUM(CASE WHEN success = 0 THEN 1 ELSE 0 END) AS failures,
+        COALESCE(SUM(input_tokens), 0) + COALESCE(SUM(output_tokens), 0) + COALESCE(SUM(cache_read_input_tokens), 0) AS totalTokens,
+        COALESCE(SUM(input_tokens), 0) AS inputTokens,
+        COALESCE(SUM(output_tokens), 0) AS outputTokens,
+        COALESCE(SUM(cache_read_input_tokens), 0) AS cacheReadTokens,
+        COALESCE(SUM(cost_usd), 0) AS costUsd
+      FROM executions
+      WHERE strftime('%Y-%m-%dT%H', started_at) >= ?
+      GROUP BY strftime('%Y-%m-%dT%H', started_at)
+    `).all(hourKeys[0]) as {
+      date: string;
+      executions: number;
+      successes: number;
+      failures: number;
+      totalTokens: number;
+      inputTokens: number;
+      outputTokens: number;
+      cacheReadTokens: number;
+      costUsd: number;
+    }[];
+
+    const byHour = new Map(rows.map((r) => [r.date, r]));
+    return hourKeys.map((date) => byHour.get(date) ?? {
+      date,
+      executions: 0,
+      successes: 0,
+      failures: 0,
+      totalTokens: 0,
+      inputTokens: 0,
+      outputTokens: 0,
+      cacheReadTokens: 0,
+      costUsd: 0,
+    });
+  }
+}
